@@ -1,23 +1,26 @@
 {
   "name": "@ali/stencil-cli",
-  "version": "2.0.0",
+  "version": "3.4.1",
   "description": "CLI tool to run BigCommerce Stores locally for theme development.",
   "main": "index.js",
   "engines": {
-    "node": ">= 10.0.0 <11.0.0"
+    "node": "^10.17 || ^12"
   },
   "scripts": {
-    "lint": "eslint .",
-    "test": "npm run lint && npm run run-tests",
-    "run-tests": "lab bin/ server/ lib/ tasks/ -t 80 -l -e 'development' -P spec",
-    "test-cov-html": "lab bin/ server/ lib/ -r html -o coverage.html -P spec"
+    "lint": "eslint . && prettier --check . --ignore-unknown",
+    "lint-and-fix": "eslint . --fix && prettier --write . --ignore-unknown",
+    "test": "jest",
+    "test-with-coverage": "jest --coverage --verbose"
   },
   "bin": {
-    "stencil": "./bin/stencil",
-    "stencil-init": "./bin/stencil-init",
-    "stencil-push": "./bin/stencil-push",
-    "stencil-start": "./bin/stencil-start",
-    "stencil-release": "./bin/stencil-release"
+    "stencil": "./bin/stencil.js",
+    "stencil-bundle": "./bin/stencil-bundle.js",
+    "stencil-download": "./bin/stencil-download.js",
+    "stencil-init": "./bin/stencil-init.js",
+    "stencil-push": "./bin/stencil-push.js",
+    "stencil-pull": "./bin/stencil-pull.js",
+    "stencil-start": "./bin/stencil-start.js",
+    "stencil-release": "./bin/stencil-release.js"
   },
   "config": {
     "stencil_version": "1.0"
@@ -32,79 +35,84 @@
     "url": "https://github.com/bigcommerce/stencil-cli/issues"
   },
   "homepage": "https://github.com/bigcommerce/stencil-cli",
+  "files": [
+    "/lib",
+    "/bin",
+    "/server",
+    "!**/*.spec.js",
+    "constants.js",
+    "package-lock.json"
+  ],
   "dependencies": {
-    "@bigcommerce/stencil-paper": "^3.0.0-rc.28",
-    "@bigcommerce/stencil-styles": "1.2.0",
-    "accept-language-parser": "^1.0.2",
-    "ajv": "^6.11.0",
-    "archiver": "^0.14.4",
-    "async": "^2.4.0",
-    "boom": "^2.7.0",
-    "browser-sync": "^2.26.7",
-    "cheerio": "^0.19.0",
+    "@bigcommerce/stencil-paper": "^3.0.0-rc.48",
+    "@bigcommerce/stencil-styles": "1.2.1",
+    "@hapi/boom": "^8.0.1",
+    "@hapi/glue": "^6.2.0",
+    "@hapi/h2o2": "^8.3.2",
+    "@hapi/hapi": "^18.4.1",
+    "@hapi/inert": "^5.2.2",
+    "@octokit/rest": "^18.0.6",
+    "ajv": "^6.12.5",
+    "archiver": "^5.0.2",
+    "async": "^3.2.0",
+    "axios": "^0.21.0",
+    "browser-sync": "^2.26.12",
+    "cheerio": "^0.22.0",
     "colors": "^1.4.0",
-    "commander": "^2.7.1",
-    "confidence": "^1.0.0",
-    "dateformat": "^2.0.0",
-    "decompress-zip": "^0.2.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^6.0.2",
-    "evil-dns": "^0.2.0",
-    "front-matter": "^1.0.0",
-    "github": "^8.1.0",
-    "glob": "^5.0.14",
-    "glue": "^2.0.0",
-    "good": "^5.1.2",
-    "good-console": "^4.1.0",
-    "graceful-fs": "^4.1.11",
-    "hapi": "^8.4.0",
-    "hoek": "^2.12.0",
-    "image-size": "^0.4.0",
-    "inquirer": "^0.8.2",
-    "jsonlint": "^1.6.2",
-    "jsonschema": "^1.0.2",
-    "jspm": "^0.16.55",
-    "lodash": "^4.17.15",
-    "memory-cache": "^0.1.4",
-    "progress": "^2.0.0",
-    "recursive-readdir": "^1.3.0",
-    "request": "^2.81.0",
-    "semver": "^5.3.0",
-    "simple-git": "^1.65.0",
-    "tarjan-graph": "^0.3.0",
-    "tmp": "0.0.26",
-    "upath": "^0.1.7",
-    "uuid4": "^1.0.0",
-    "when": "^3.7.7",
-    "wreck": "^12.2.2",
+    "commander": "^6.1.0",
+    "confidence": "^4.0.2",
+    "form-data": "^3.0.0",
+    "front-matter": "^4.0.2",
+    "glob": "^7.1.6",
+    "graceful-fs": "^4.2.4",
+    "image-size": "^0.9.1",
+    "inquirer": "^7.3.3",
+    "jsonlint": "^1.6.3",
+    "lodash": "^4.17.20",
+    "memory-cache": "^0.2.0",
+    "npm-which": "^3.0.1",
+    "object-to-spawn-args": "^2.0.0",
+    "progress": "^2.0.3",
+    "recursive-readdir": "^2.2.2",
+    "semver": "^7.3.2",
+    "simple-git": "^2.20.1",
+    "tarjan-graph": "^2.0.0",
+    "tmp-promise": "^3.0.2",
+    "upath": "^1.2.0",
+    "uuid4": "^2.0.2",
     "yauzl": "^2.10.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^7.2.1",
-    "@commitlint/config-conventional": "^7.1.2",
-    "babel-core": "^6.8.0",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
-    "code": "^4.1.0",
-    "conventional-changelog-cli": "^1.3.2",
+    "@commitlint/cli": "^11.0.0",
+    "@commitlint/config-conventional": "^11.0.0",
+    "axios-mock-adapter": "^1.19.0",
+    "conventional-changelog-cli": "^2.1.0",
+    "eslint": "^7.10.0",
+    "eslint-config-airbnb-base": "^14.2.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-jest": "^24.0.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.1.4",
     "gulp": "^4.0.2",
-    "gulp-bump": "^3.1.3",
-    "gulp-eslint": "^6.0.0",
-    "gulp-exec": "^4.0.0",
-    "gulp-git-streamed": "^2.8.1",
+    "gulp-bump": "^3.2.0",
+    "gulp-exec": "^5.0.0",
+    "gulp-git-streamed": "^2.10.1",
     "gulp-if": "^3.0.0",
     "gulp-prompt": "^1.2.0",
     "gulp-util": "^3.0.8",
-    "husky": "^4.2.1",
-    "jest": "^21.2.1",
-    "lab": "^14.3.3",
-    "npm-which": "^3.0.1",
-    "object-to-spawn-args": "^1.1.1",
-    "sinon": "^8.1.1"
+    "husky": "^4.3.0",
+    "jest": "^26.4.2",
+    "lint-staged": "^10.4.0",
+    "prettier": "2.1.2"
   },
   "husky": {
     "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
+      "pre-commit": "lint-staged"
     }
+  },
+  "lint-staged": {
+    "*": "npm run lint-and-fix"
   }
 }
