@@ -1,11 +1,9 @@
-const _ = require('lodash');
-const Async = require('async');
-const Fs = require('fs');
-const Path = require('path');
-const Upath = require('upath');
-const importRegex = /@import\s+?["'](.+?)["'];/g;
+const async = require('async');
+const fs = require('fs');
+const path = require('path');
+const upath = require('upath');
 
-module.exports.assemble = assemble;
+const importRegex = /@import\s+?["'](.+?)["'];/g;
 
 /**
  * Parses all of the imports for SCSS files and returns a flat object with the file path as the key and
@@ -18,101 +16,81 @@ module.exports.assemble = assemble;
  * @param callback
  */
 function assemble(cssFiles, absolutePath, type, options, callback) {
-    var ret = {},
-        ext = '.' + type;
+    const ret = {};
+    const ext = `.${type}`;
 
-    if (typeof options === 'function') {
-        callback = options;
-        options = {};
-    }
+    const cssFilesArr = Array.isArray(cssFiles) ? cssFiles : [cssFiles];
 
-    if (! _.isArray(cssFiles)) {
-        cssFiles = [cssFiles];
-    }
+    const parseImports = (cssFileAlias, parseImportsCallback) => {
+        const normalizedCssFileAlias = cssFileAlias.replace(/(\.scss)$/g, '');
+        const cssFilePath = path.join(absolutePath, normalizedCssFileAlias + ext);
+        const cssFileAliasDir = path.parse(normalizedCssFileAlias).dir;
 
-    Async.map(cssFiles, function(cssFile, mapCallback) {
-        var cssFileParts = Path.parse(cssFile),
-            cssFileAlias = Path.join(cssFileParts.dir, cssFileParts.name);
+        const fileParts = path.parse(cssFilePath);
+        const underscoredFilePath = path.join(fileParts.dir, `_${fileParts.base}`);
+        let filePath;
 
-        parseImports(cssFileAlias, mapCallback);
-    }, function (err) {
-        if (err) {
-            return callback(err);
+        if (fs.existsSync(cssFilePath)) {
+            filePath = cssFilePath;
+        } else if (fs.existsSync(underscoredFilePath)) {
+            filePath = underscoredFilePath;
+        } else {
+            // File doesn't exist, just return to skip it
+            parseImportsCallback();
+            return;
         }
 
-        callback(null, ret);
-    });
+        fs.readFile(filePath, { encoding: 'utf-8' }, (err, content) => {
+            const matches = [];
+            let match;
+            let importCssFileAlias;
 
-    function parseImports(cssFileAlias, callback) {
-        cssFileAlias = cssFileAlias.replace(/(\.scss)$/g, '');
-        var file = Path.join(absolutePath, cssFileAlias + ext),
-            cssFileAliasDir = Path.parse(cssFileAlias).dir;
+            if (!err) {
+                // Ensure all import paths are Unix paths for prod compat.
+                const cssFile = options.bundle
+                    ? upath.toUnix(normalizedCssFileAlias + ext)
+                    : normalizedCssFileAlias + ext;
 
-        Async.parallel({
-            originalFileExists: function(callback) {
-                Fs.exists(file, function(exists) {
-                    callback(null, exists ? file : false);
-                });
-            },
-            underscoredFileExists: function(callback) {
-                var fileParts = Path.parse(file),
-                    underscoredFile = Path.join(fileParts.dir, '_' + fileParts.base);
+                ret[cssFile] = content;
+                match = importRegex.exec(content);
 
-                Fs.exists(underscoredFile, function(exists) {
-                    callback(null, exists ? underscoredFile : false);
-                });
-            },
-        },
-        function(err, results) {
-            var file = false;
-
-            if (results.originalFileExists) {
-                file = results.originalFileExists;
-            }
-
-            if (results.underscoredFileExists) {
-                file = results.underscoredFileExists;
-            }
-
-            // File doesn't exist, just return to skip it
-            if (! file) {
-                return callback();
-            }
-
-            Fs.readFile(file, {encoding: 'utf-8'}, function(err, content) {
-                var matches = [],
-                    match,
-                    importCssFileAlias,
-                    cssFile;
-
-                if (! err) {
-
-                    //Ensure all import paths are Unix paths for prod compat.
-                    if (options.bundle) {
-                        cssFile = Upath.toUnix(cssFileAlias + ext);
-                    } else {
-                        cssFile = cssFileAlias + ext;
+                while (match !== null) {
+                    [, importCssFileAlias] = match;
+                    if (cssFileAliasDir) {
+                        importCssFileAlias = path.join(cssFileAliasDir, importCssFileAlias);
                     }
 
-                    ret[cssFile] = content;
+                    if (!ret[importCssFileAlias + ext]) {
+                        matches.push(importCssFileAlias);
+                    }
+
                     match = importRegex.exec(content);
-
-                    while (match !== null) {
-                        importCssFileAlias = match[1];
-                        if (cssFileAliasDir) {
-                            importCssFileAlias = Path.join(cssFileAliasDir, importCssFileAlias);
-                        }
-
-                        if (! ret[importCssFileAlias + ext]) {
-                            matches.push(importCssFileAlias);
-                        }
-
-                        match = importRegex.exec(content);
-                    }
                 }
+            }
 
-                Async.each(matches, parseImports, callback);
-            });
+            async.each(matches, parseImports, parseImportsCallback);
         });
-    }
+    };
+
+    async.map(
+        cssFilesArr,
+        (cssFile, mapCallback) => {
+            const cssFileParts = path.parse(cssFile);
+            const cssFileAlias = path.join(cssFileParts.dir, cssFileParts.name);
+
+            parseImports(cssFileAlias, mapCallback);
+        },
+        (err) => {
+            if (err) {
+                callback(err);
+                return;
+            }
+
+            callback(null, ret);
+        },
+    );
 }
+
+module.exports = {
+    assemble,
+};
