@@ -1,13 +1,11 @@
-var jsonLint = require('jsonlint');
+const jsonLint = require('jsonlint');
 
 module.exports = {
-    parse: function (jsonString, file) {
-        file = file || '';
-
+    parse(jsonString, file = '') {
         try {
             return jsonLint.parse(jsonString);
         } catch (e) {
-            throw new Error(file + ' - ' + e.message);
+            throw new Error(`${file} - ${e.message}`);
         }
     },
 };
