@@ -1,32 +1,23 @@
-var Code = require('code'),
-    Lab = require('lab'),
-    lab = exports.lab = Lab.script(),
-    describe = lab.describe,
-    jsonLint = require('./json-lint'),
-    expect = Code.expect,
-    it = lab.it;
+const jsonLint = require('./json-lint');
 
-describe('json-lint', function () {
-    var badJsonFilename = '/path/to/badfile.json',
-        badJson = '{"foo":"bar" "fizz": "buzz"}',
-        file = new RegExp(badJsonFilename);
+describe('json-lint', () => {
+    const badJsonFilename = '/path/to/badfile.json';
+    const badJson = '{"foo":"bar" "fizz": "buzz"}';
+    const file = new RegExp(badJsonFilename);
 
-
-    it('should add file name to error', function (done) {
-        var throws = function () {
+    it('should add file name to error', () => {
+        const throws = () => {
             jsonLint.parse(badJson, badJsonFilename);
         };
 
-        expect(throws).throw(Error, file);
-        done();
+        expect(throws).toThrow(Error, file);
     });
 
-    it('should not need a file name', function (done) {
-        var throws = function () {
+    it('should not need a file name', () => {
+        const throws = () => {
             jsonLint.parse(badJson);
         };
 
-        expect(throws).throw(Error, !file);
-        done();
+        expect(throws).toThrow(Error, !file);
     });
 });
