@@ -1,11 +1,8 @@
-var Async = require('async'),
-    Fs = require('fs'),
-    Path = require('path'),
-    _ = require('lodash'),
-    localeDirectory = 'lang';
+const async = require('async');
+const fs = require('fs');
+const path = require('path');
 
-module.exports.assemble = assemble;
-module.exports.localeDirectory = localeDirectory;
+const LOCALE_DIRECTORY = 'lang';
 
 /**
  * Assembles together all of the lang files.
@@ -15,29 +12,32 @@ module.exports.localeDirectory = localeDirectory;
  * @param callback
  */
 function assemble(callback) {
-    Fs.readdir(localeDirectory, function (err, localeFiles) {
-        var localesToLoad = {};
+    fs.readdir(LOCALE_DIRECTORY, (err, localeFiles) => {
+        if (err) {
+            callback(err);
+            return;
+        }
 
         // Ignore hidden files
         // @example: MAC generates .DS_STORE
-        localeFiles = _.filter(localeFiles, function(fileName) {
-            return fileName[0] !== '.';
-        });
+        const filteredLocaleFiles = localeFiles.filter((fileName) => fileName[0] !== '.');
+        const localesToLoad = {};
 
-        if (err) {
-            return callback(err);
+        for (const localeFile of filteredLocaleFiles) {
+            const localeName = path.basename(localeFile, '.json');
+
+            localesToLoad[localeName] = (cb2) => {
+                const localeFilePath = `${LOCALE_DIRECTORY}/${localeFile}`;
+
+                fs.readFile(localeFilePath, 'utf-8', cb2);
+            };
         }
 
-        _.each(localeFiles, function (localeFile) {
-            var localeName = Path.basename(localeFile, '.json');
-
-            localesToLoad[localeName] = function (callback) {
-                var localeFilePath = localeDirectory + '/' + localeFile;
-
-                Fs.readFile(localeFilePath, 'utf-8', callback);
-            };
-        });
-
-        Async.parallel(localesToLoad, callback);
+        async.parallel(localesToLoad, callback);
     });
 }
+
+module.exports = {
+    assemble,
+    LOCALE_DIRECTORY,
+};
