@@ -1,119 +1,47 @@
-'use strict';
-
 const os = require('os');
 const uuid = require('uuid4');
 const fs = require('fs');
 const path = require('path');
-const themePath = process.cwd();
-const git = require('simple-git')(themePath);
-const themeConfig = require('../theme-config').getInstance(themePath);
-const questions = require('./questions');
-const GitHub = require('github');
+const { Octokit } = require('@octokit/rest');
+const simpleGit = require('simple-git');
+
+const ThemeConfig = require('../theme-config');
+const askQuestions = require('./questions');
 const Bundle = require('../stencil-bundle');
+const StencilConfigManager = require('../StencilConfigManager');
+const { parseJsonFile } = require('../utils/fsUtils');
+const { THEME_PATH } = require('../../constants');
 
-module.exports = () => {
+const git = simpleGit(THEME_PATH);
+const themeConfigManager = ThemeConfig.getInstance(THEME_PATH);
+const stencilConfigManager = new StencilConfigManager();
 
-    getGitData((err, gitData) => {
-        if (err) {
-            throw err;
-        }
+async function saveGithubToken(githubToken) {
+    const data = (await stencilConfigManager.read(true)) || {};
 
-        if (gitData.remotes.length === 0) {
-            return printError('No git remote repository found');
-        }
+    data.githubToken = githubToken;
 
-        if (gitData.current !== 'master') {
-            return printError('Not on master branch, please checkout master branch to proceed');
-        }
+    await stencilConfigManager.save(data);
+}
 
-        if (gitData.dirty) {
-            return printError('git tree is dirty, please commit changes to proceed');
-        }
+async function getGithubToken() {
+    const data = (await stencilConfigManager.read(true)) || {};
 
-        if (gitData.behind !== 0) {
-            return printError(`Your branch is behind by ${gitData.behind} commits`);
-        }
+    return data.githubToken;
+}
 
-        if (gitData.ahead !== 0) {
-            printWarning(`your branch is ahead by ${gitData.ahead} commits`);
-        }
+function isReleaseCandidate(version) {
+    return version.includes('-rc.');
+}
 
-        questions(themeConfig, getGithubToken(), gitData.remotes, (err, answers) => {
-            if (err) {
-                return printError(err.message);
-            }
-
-            saveGithubToken(answers.githubToken);
-
-            doRelease(answers, err => {
-                if (err) {
-                    return printError(err.message);
-                }
-
-                console.log('done'.green);
-            });
-        });
-    });
-};
-
-function doRelease(options, callback) {
-    // Update changelog and get text for release notes
-    const changelog = parseChangelog(options.version, options.date);
-
-    bumpJsonFileVersion(path.join(themePath, 'config.json'), options.version);
-    bumpJsonFileVersion(path.join(themePath, 'package.json'), options.version);
-
-    bundleTheme((err, bundlePath) => {
-        if (err) {
-            return callback(err);
-        }
-
-        commitAndPush(options.version, options.remote, (err, commit) => {
-            if (err) {
-                fs.unlinkSync(bundlePath);
-                return callback(err);
-            }
-
-            createRelease(options, commit, bundlePath, changelog, err => {
-                fs.unlinkSync(bundlePath);
-                if (err) {
-                    return callback(err);
-                }
-
-                callback();
-            });
-        });
+async function getGithubClient() {
+    return new Octokit({
+        auth: await getGithubToken(),
     });
 }
 
-function commitAndPush(version, remote, callback) {
-    git.add(['config.json', 'package.json', 'CHANGELOG.md']).commit(`Releasing ${version}`, (err, summary) => {
-        if (err) {
-            return callback(err);
-        }
-
-        console.log(`Pushing Changes to ${remote.name}...`);
-
-        git.push(remote.name, 'master', err => {
-            if (err) {
-                return callback(err);
-            }
-
-            callback(null, summary.commit);
-        });
-    });
-}
-
-function createRelease(options, commit, bundlePath, changelog, callback) {
-    if (options.createGithubRelease) {
-        createGithubRelease(commit, options.version, changelog, options.remote, bundlePath, callback);
-    } else {
-        process.nextTick(callback);
-    }
-}
-
-function createGithubRelease(commit, version, changelog, remote, bundlePath, callback) {
-    const github = getGithubClient();
+async function createGithubRelease(commit, version, changelog, remote, bundlePath) {
+    const github = await getGithubClient();
 
     const releaseParams = {
         owner: remote.owner,
@@ -125,43 +53,38 @@ function createGithubRelease(commit, version, changelog, remote, bundlePath, cal
 
     console.log('Creating Github Release...');
 
-    github.repos.createRelease(releaseParams)
-        .then(release => {
-            const uploadParams = {
-                id: release.id,
-                owner: remote.owner,
-                repo: remote.repo,
-                filePath: bundlePath,
-                name: `${themeConfig.getName()}-${version}.zip`,
-            };
+    const release = await github.repos.createRelease(releaseParams);
+    const themeName = await themeConfigManager.getName();
 
-            console.log('Uploading Bundle File...');
+    console.log('Uploading Bundle File...');
 
-            github.repos.uploadAsset(uploadParams)
-                .then(asset => {
-                    console.log(`Release url: ${release.html_url.green}`);
-                    console.log(`Bundle download url: ${asset.browser_download_url.green}`);
+    const zipFile = await fs.promises.readFile(bundlePath);
+    const uploadParams = {
+        release_id: release.data.id,
+        owner: remote.owner,
+        repo: remote.repo,
+        data: zipFile,
+        name: `${themeName}-${version}.zip`,
+    };
+    const asset = await github.repos.uploadReleaseAsset(uploadParams);
 
-                    callback();
-                })
-                .catch(callback);
-        })
-        .catch(callback);
+    console.log(`Release url: ${release.data.html_url.green}`);
+    console.log(`Bundle download url: ${asset.data.browser_download_url.green}`);
 }
 
-function bumpJsonFileVersion(filePath, version) {
-    const data = JSON.parse(fs.readFileSync(filePath));
+async function bumpJsonFileVersion(filePath, version) {
+    const data = await parseJsonFile(filePath);
 
     data.version = version;
-    fs.writeFileSync(filePath, JSON.stringify(data, null, 2) + '\n');
+    await fs.promises.writeFile(filePath, `${JSON.stringify(data, null, 2)}\n`);
 }
 
-function parseChangelog(version, date) {
-    const filePath = path.join(themePath, 'CHANGELOG.md');
-    var changelog = '';
+async function parseChangelog(version, date) {
+    const filePath = path.join(THEME_PATH, 'CHANGELOG.md');
+    let changelog = '';
 
     try {
-        changelog = fs.readFileSync(filePath).toString();
+        changelog = await fs.promises.readFile(filePath, 'utf8');
     } catch (e) {
         // no CHANGELOG file
         return null;
@@ -174,81 +97,49 @@ function parseChangelog(version, date) {
 
     // Update the changelog file if this is a full release
     if (!isReleaseCandidate(version)) {
-        fs.writeFileSync(filePath, changelog);
+        await fs.promises.writeFile(filePath, changelog);
     }
 
     return match ? match[1].trim() : null;
 }
 
-function saveGithubToken(githubToken) {
-    const dotStencilPath = path.join(themePath, '.stencil');
-    var data = {};
-
-    try {
-        data = JSON.parse(fs.readFileSync(dotStencilPath));
-    } catch (e) {
-        // .stencil file might not exist
-    }
-
-    data.githubToken = githubToken;
-    fs.writeFileSync(dotStencilPath, JSON.stringify(data, null, 2) + '\n');
-}
-
-function getGithubToken() {
-    const dotStencilPath = path.join(themePath, '.stencil');
-    var data = {};
-
-    try {
-        data = JSON.parse(fs.readFileSync(dotStencilPath));
-    } catch (e) {
-        // .stencil file might not exist
-    }
-
-    return data.githubToken;
-}
-
-function bundleTheme(callback) {
-    const bundle = new Bundle(themePath, themeConfig, themeConfig.getRawConfig(), {
+async function bundleTheme() {
+    const rawConfig = await themeConfigManager.getRawConfig();
+    const bundleOptions = {
         dest: os.tmpdir(),
         name: uuid(),
-    });
+    };
+    const bundle = new Bundle(THEME_PATH, themeConfigManager, rawConfig, bundleOptions);
 
-    bundle.initBundle(callback);
+    return bundle.initBundle();
 }
 
-function getGitData(callback) {
-    git.status((err, status) => {
-        if (err) {
-            return callback(err);
-        }
+async function getGitData() {
+    const status = await git.status();
 
-        git.getRemotes(true, (err, response) => {
-            if (err) {
-                return callback(err);
-            }
+    const response = await git.getRemotes(true);
 
-            const remotes = response.map(remote => {
-                const url = remote.refs.push || '';
-                const match = url.match(/github\.com[\/|:](.+?)\/(.+?)[\/|\.]/);
+    const remotes = response.map((remote) => {
+        const url = remote.refs.push || '';
+        const match = url.match(/github\.com[/|:](.+?)\/(.+?)[/|.]/);
 
-                return {
-                    name: remote.name,
-                    url,
-                    owner: match ? match[1] : null,
-                    repo: match ? match[2] : null,
-                };
-            });
-
-            const dirty = status.not_added.length > 0 ||
-                status.conflicted.length > 0 ||
-                status.created.length > 0 ||
-                status.deleted.length > 0 ||
-                status.modified.length > 0 ||
-                status.renamed.length > 0;
-
-            callback(null, Object.assign({ dirty }, status, { remotes }));
-        });
+        return {
+            name: remote.name,
+            url,
+            owner: match ? match[1] : null,
+            repo: match ? match[2] : null,
+        };
     });
+
+    const dirty =
+        status.not_added.length > 0 ||
+        status.conflicted.length > 0 ||
+        status.created.length > 0 ||
+        status.deleted.length > 0 ||
+        status.modified.length > 0 ||
+        status.renamed.length > 0;
+
+    return { dirty, ...status, remotes };
 }
 
 function printError(message) {
@@ -256,20 +147,106 @@ function printError(message) {
 }
 
 function printWarning(message) {
-    console.log('\nWarning: '.yellow + `${message}\n`);
+    console.log('\nWarning: '.yellow + message);
 }
 
-function getGithubClient() {
-    const github = new GitHub();
+/**
+ * @param gitData
+ * @param gitData.remotes
+ * @param gitData.current
+ * @param gitData.dirty
+ * @param gitData.behind
+ * @param gitData.ahead
+ * @returns {boolean}
+ */
+function checkGitData(gitData) {
+    if (gitData.remotes.length === 0) {
+        printError('No git remote repository found');
+        return false;
+    }
 
-    github.authenticate({
-        type: 'oauth',
-        token: getGithubToken(),
-    });
+    if (gitData.current !== 'master') {
+        printError('Not on master branch, please checkout master branch to proceed');
+        return false;
+    }
 
-    return github;
+    if (gitData.dirty) {
+        printError('git tree is dirty, please commit changes to proceed');
+        return false;
+    }
+
+    if (gitData.behind !== 0) {
+        printError(`Your branch is behind by ${gitData.behind} commits`);
+        return false;
+    }
+
+    if (gitData.ahead !== 0) {
+        printWarning(`your branch is ahead by ${gitData.ahead} commits`);
+    }
+
+    return true;
 }
 
-function isReleaseCandidate(version) {
-    return version.includes('-rc.');
+async function commitAndPush(version, remote) {
+    await git.add(['config.json', 'package.json', 'package-lock.json', 'CHANGELOG.md']);
+    const summary = await git.commit(`Releasing ${version}`);
+
+    console.log(`Pushing Changes to ${remote.name}...`);
+
+    await git.push(remote.name, 'master');
+
+    return summary.commit;
 }
+
+async function doRelease(options) {
+    // Update changelog and get text for release notes
+    const changelog = await parseChangelog(options.version, options.date);
+
+    await bumpJsonFileVersion(themeConfigManager.configPath, options.version);
+    await bumpJsonFileVersion(path.join(THEME_PATH, 'package.json'), options.version);
+    await bumpJsonFileVersion(path.join(THEME_PATH, 'package-lock.json'), options.version);
+
+    const bundlePath = await bundleTheme();
+
+    try {
+        const commit = await commitAndPush(options.version, options.remote);
+
+        if (options.createGithubRelease) {
+            await createGithubRelease(
+                commit,
+                options.version,
+                changelog,
+                options.remote,
+                bundlePath,
+            );
+        }
+    } finally {
+        await fs.promises.unlink(bundlePath);
+    }
+}
+
+async function run() {
+    const gitData = await getGitData();
+
+    if (!checkGitData(gitData)) {
+        return;
+    }
+
+    try {
+        const answers = await askQuestions(
+            themeConfigManager,
+            await getGithubToken(),
+            gitData.remotes,
+        );
+
+        await saveGithubToken(answers.githubToken);
+
+        await doRelease(answers);
+
+        console.log('done'.green);
+    } catch (err) {
+        printError(err.message || err);
+    }
+}
+
+module.exports = run;
