@@ -1,431 +1,443 @@
-'use strict';
+const upath = require('upath');
+const Archiver = require('archiver');
+const async = require('async');
+const crypto = require('crypto');
+const fs = require('fs');
+const path = require('path');
 
-const MEGABYTE = 1048576;
+const BuildConfigManager = require('./BuildConfigManager');
+const BundleValidator = require('./bundle-validator');
+const Cycles = require('./Cycles');
+const cssAssembler = require('./css-assembler');
+const langAssembler = require('./lang-assembler');
+const templateAssembler = require('./template-assembler');
+const { recursiveReadDir } = require('./utils/fsUtils');
+const { fetchRegions } = require('./regions');
+
+const MEGABYTE = 1024 * 1024;
 const MAX_SIZE_BUNDLE = MEGABYTE * 50;
 const PATHS_TO_ZIP = [
-    'assets/**/*',
-    '!assets/cdn/**',
-    '!assets/**/*.js.map',
-    '!assets/jspm_packages/**', // Don't want jspm_packages if it's there
-    'CHANGELOG.md',
-    'config.json',
-    '.eslintrc',
-    '.eslintignore',
-    'Gruntfile.js',
-    'karma.conf.js',
-    'lang/*',
-    'meta/**/*',
-    'package.json',
-    'README.md',
-    '.scss-lint.yml',
-    'stencil.conf.js',
-    'templates/**/*',
-    'webpack.*.js',
+    {
+        pattern: 'assets/**/*',
+        ignore: ['assets/cdn/**', 'assets/**/*.js.map'],
+    },
+    { pattern: 'CHANGELOG.md' },
+    { pattern: 'config.json' },
+    { pattern: '.eslintrc' },
+    { pattern: '.eslintignore' },
+    { pattern: 'Gruntfile.js' },
+    { pattern: 'karma.conf.js' },
+    { pattern: 'lang/*' },
+    { pattern: 'meta/**/*' },
+    { pattern: 'package.json' },
+    { pattern: 'README.md' },
+    { pattern: '.scss-lint.yml' },
+    { pattern: 'stencil.conf.js' },
+    { pattern: 'templates/**/*' },
+    { pattern: 'webpack.*.js' },
 ];
 
-const Upath = require('upath');
-const Tmp = require('tmp');
-const Readdir = require('recursive-readdir');
-const Archiver = require('archiver');
-const Async = require('async');
-const Crypto = require('crypto');
-const Fs = require('fs');
-const Jspm = require('jspm');
-const Path = require('path');
-const buildConfig = require('../lib/build-config');
-const BundleValidator = require('./bundle-validator');
-const Cycles = require('./cycles');
-const CssAssembler = require('./css-assembler');
-const LangAssembler = require('./lang-assembler');
-const TemplateAssembler = require('./template-assembler');
-const Regions = require('./regions');
+class Bundle {
+    constructor(
+        themePath,
+        themeConfig,
+        rawConfig,
+        options = {},
+        buildConfigManger = new BuildConfigManager(),
+    ) {
+        const tasks = {};
+        this.options = options;
 
-function Bundle(themePath, themeConfig, rawConfig, options) {
-    const tasks = {};
-    this.options = options || {};
+        this.templatesPath = path.join(themePath, 'templates');
+        this.themePath = themePath;
+        this.themeConfig = themeConfig;
+        this.configuration = rawConfig;
 
-    this.templatesPath = Path.join(themePath, 'templates');
-    this.themePath = themePath;
-    this.themeConfig = themeConfig;
-    this.configuration = rawConfig;
+        this.validator = new BundleValidator(
+            this.themePath,
+            this.themeConfig,
+            this.options.marketplace !== true,
+        );
 
-    this.validator = new BundleValidator(this.themePath, this.themeConfig, this.options.marketplace !== true);
-
-    if (this.configuration.css_compiler) {
-        tasks.css = this.getCssAssembleTask(this.configuration.css_compiler);
-    }
-
-    tasks.templates = this.assembleTemplatesTask.bind(this);
-    tasks.lang = this.assembleLangTask.bind(this);
-    tasks.schema = this.assembleSchema.bind(this);
-    tasks.schemaTranslations = this.assembleSchemaTranslations.bind(this);
-
-    if (typeof buildConfig.production === 'function') {
-        tasks.theme = callback => {
-            console.log('Theme task Started...');
-            buildConfig.initWorker().production(err => {
-                if (err) {
-                    return callback(err);
-                }
-
-                console.log('ok'.green + ' -- Theme task Finished');
-                callback();
-            });
-        };
-    }
-
-    this.tasks = tasks;
-
-    if (this.configuration.jspm) {
-        this.configuration.jspm.tmpBundleFile = Tmp.fileSync().name;
-        this.tasks.jspmBundle = this.getJspmBundleTask(this.configuration.jspm);
-    }
-}
-
-/**
- * Initializes bundling process and executes all required tasks.
- * @param {Function} callback
- */
-Bundle.prototype.initBundle = function (callback) {
-    Async.series({
-        validate: validateTheme.bind(this),
-        bundle: bundleTaskRunner.bind(this),
-    }, (err, result) => {
-        let errorMessage = '';
-
-        if (err) {
-            errorMessage = err.message ? err.message : String(err);
-            console.error('failed  -- '.red + errorMessage.red);
-            return callback(err);
+        if (this.configuration.css_compiler) {
+            tasks.css = this.getCssAssembleTask(this.configuration.css_compiler);
         }
 
-        callback(null, result.bundle);
-    });
-};
+        tasks.templates = this.assembleTemplatesTask.bind(this);
+        tasks.lang = this.assembleLangTask.bind(this);
+        tasks.schema = this.assembleSchema.bind(this);
+        tasks.schemaTranslations = this.assembleSchemaTranslations.bind(this);
 
-Bundle.prototype.getCssAssembleTask = function (compiler) {
-    const assembleOptions = {
-        bundle: true,
-    };
-    return callback => {
-        const basePath = Path.join(this.themePath, 'assets', compiler);
+        if (typeof buildConfigManger.production === 'function') {
+            tasks.theme = (callback) => {
+                console.log('Theme task Started...');
+                buildConfigManger.initWorker().production((err) => {
+                    if (err) {
+                        return callback(err);
+                    }
 
-        console.log('%s Parsing Started...', compiler.toUpperCase());
-        Fs.readdir(basePath, (err, files) => {
-            const filterFiles = files.filter(file => {
-                return file.substr(-(compiler.length + 1)) === '.' + compiler;
-            });
-            Async.map(filterFiles, (file, mapCallback) => {
-                CssAssembler.assemble(file, basePath, compiler, assembleOptions, mapCallback);
-            }, (err, results) => {
-                const ret = {};
-                if (err) {
-                    return callback(err);
-                }
+                    console.log(`${'ok'.green} -- Theme task Finished`);
 
-                filterFiles.forEach((file, index) => {
-                    ret[file] = results[index];
+                    return callback();
                 });
-
-                console.log('ok'.green + ' -- %s Parsing Finished', compiler.toUpperCase());
-                callback(null, ret);
-            });
-        });
-    };
-};
-
-Bundle.prototype.assembleTemplatesTask = function (callback) {
-    console.log('Template Parsing Started...');
-
-    Readdir(this.templatesPath, ['!*.html'], (err, files) => {
-        if (err) {
-            return callback(err);
+            };
         }
 
-        const partials = files.map(file => {
-            return Upath.toUnix(file.replace(this.templatesPath + Path.sep, '').replace(/\.html$/, ''));
-        });
-
-        Async.map(partials, TemplateAssembler.assembleAndBundle.bind(null, this.templatesPath), (err, results) => {
-            const ret = {};
-
-            if (err) {
-                return callback(err);
-            }
-
-            partials.forEach((file, index) => {
-                ret[file] = results[index];
-            });
-
-            Async.parallel([
-                checkObjects.bind(this, results),
-                detectCycles.bind(this, results),
-            ], err => {
-                if (err) {
-                    callback(err);
-                }
-
-                console.log('ok'.green + ' -- Template Parsing Finished');
-                callback(null, ret);
-            });
-        });
-    });
-};
-
-Bundle.prototype.assembleSchema = function (callback) {
-    console.log('Building Theme Schema File...');
-
-    this.themeConfig.getSchema((err, schema) => {
-        if (err) {
-            callback(err);
-        }
-
-        console.log('ok'.green + ' -- Theme Schema Building Finished');
-
-        callback(null, schema);
-    });
-};
-
-Bundle.prototype.assembleSchemaTranslations = function (callback) {
-    console.log('Schema Translations Parsing Started...');
-
-    this.themeConfig.getSchemaTranslations((err, schema) => {
-        if (err) {
-            callback(err);
-        }
-
-        console.log('ok'.green + ' -- Schema Translations Parsing Finished');
-
-        callback(null, schema);
-    });
-};
-
-Bundle.prototype.assembleLangTask = function (callback) {
-    console.log('Language Files Parsing Started...');
-    LangAssembler.assemble((err, results) => {
-        if (err) {
-            return callback(err);
-        }
-
-        console.log('ok'.green + ' -- Language Files Parsing Finished');
-        callback(null, results);
-    });
-};
-
-Bundle.prototype.getJspmBundleTask = function (jspmConfig) {
-    return callback => {
-        const oldConsoleError = console.error;
-
-        console.log('JavaScript Bundling Started...');
-
-        // Need to suppress annoying errors from Babel.
-        // They will be gone in the next minor version of JSPM (0.16.0).
-        // Until then, this will stay in place
-        console.error = error => {
-            if (!/Deprecated option metadataUsedHelpers/.test(error)) {
-                oldConsoleError(error);
-            }
-        };
-
-        Jspm.setPackagePath(this.themePath);
-        Jspm.bundleSFX(jspmConfig.bootstrap, jspmConfig.tmpBundleFile, {
-            minify: true,
-            mangle: true,
-        }).then(() => {
-            console.log('ok'.green + ' -- JavaScript Bundling Finished');
-            console.error = oldConsoleError;
-            callback(null, true);
-        }).catch(err => {
-            callback(err);
-        });
-    };
-};
-
-Bundle.prototype.generateManifest = function (taskResults, callback) {
-    console.log('Generating Manifest Started...');
-
-    Readdir(this.templatesPath, ['!*.html'], (err, filePaths) => {
-        if (err) {
-            return callback(err);
-        }
-
-        const templates = filePaths.map(file => {
-            return Upath.toUnix(file.replace(this.templatesPath + Path.sep, '').replace('.html', ''));
-        });
-
-        const regions = Regions.fetchRegions(taskResults.templates, templates);
-
-        console.log('ok'.green + ' -- Manifest Generation Finished');
-        return callback(null, {
-            regions,
-            templates,
-        });
-    });
-};
-
-function checkObjects(results, callback) {
-    this.validator.validateObjects(results, err => {
-        if (err) {
-            console.error('error '.red + err.message);
-            return callback(err);
-        }
-
-        callback();
-    });
-}
-
-function detectCycles(results, callback) {
-    try {
-        new Cycles(results).detect();
-        callback();
-    } catch (err) {
-        callback(err);
+        this.tasks = tasks;
     }
-}
 
-function validateTheme(callback) {
-    console.log("Validating theme...");
-    this.validator.validateTheme(err => {
-        if (err) {
+    /**
+     * Initializes bundling process and executes all required tasks.
+     *
+     * @returns {Promise<string>} - path of the generated bundle
+     */
+    async initBundle() {
+        try {
+            const result = await async.series({
+                validate: this._validateTheme.bind(this),
+                bundle: this._bundleTaskRunner.bind(this),
+            });
+            return result.bundle;
+        } catch (err) {
+            const errorMessage = err.message ? err.message : String(err);
+            console.error('failed  -- '.red + errorMessage.red);
             throw err;
         }
-
-        callback(null, true);
-    });
-}
-
-function bundleTaskRunner(callback) {
-    let defaultName = this.configuration.name + '-' + this.configuration.version + '.zip';
-    const outputName = typeof this.options.name === 'string' ? this.options.name + '.zip' : defaultName;
-    const outputFolder = typeof this.options.dest === 'string' ? this.options.dest : this.themePath;
-    const bundleZipPath = Path.join(outputFolder, outputName);
-
-    if (!this.configuration.name) {
-        defaultName = 'Theme.zip';
     }
 
-    Async.parallel(this.tasks, (err, taskResults) => {
-        if (err) {
-            return callback(err);
-        }
+    getCssAssembleTask(compiler) {
+        const assembleOptions = {
+            bundle: true,
+        };
+        return (callback) => {
+            const basePath = path.join(this.themePath, 'assets', compiler);
 
-        const archive = Archiver('zip');
-        const fileStream = Fs.createWriteStream(bundleZipPath);
-        archive.pipe(fileStream);
+            console.log('%s Parsing Started...', compiler.toUpperCase());
+            fs.readdir(basePath, (err, files) => {
+                const filterFiles = files.filter((file) => {
+                    return file.substr(-(compiler.length + 1)) === `.${compiler}`;
+                });
+                async.map(
+                    filterFiles,
+                    (file, mapCallback) => {
+                        cssAssembler.assemble(
+                            file,
+                            basePath,
+                            compiler,
+                            assembleOptions,
+                            mapCallback,
+                        );
+                    },
+                    (assemblingError, results) => {
+                        const ret = {};
+                        if (assemblingError) {
+                            return callback(assemblingError);
+                        }
 
-        // Create manifest will use taskResults to generate a manifest file
-        this.generateManifest(taskResults, (err, manifest) => {
+                        filterFiles.forEach((file, index) => {
+                            ret[file] = results[index];
+                        });
+
+                        console.log(`${'ok'.green} -- %s Parsing Finished`, compiler.toUpperCase());
+                        return callback(null, ret);
+                    },
+                );
+            });
+        };
+    }
+
+    assembleTemplatesTask(callback) {
+        console.log('Template Parsing Started...');
+
+        recursiveReadDir(this.templatesPath, ['!*.html'], (readdirError, files) => {
+            if (readdirError) {
+                return callback(readdirError);
+            }
+
+            const partials = files.map((file) => {
+                return upath.toUnix(
+                    file.replace(this.templatesPath + path.sep, '').replace(/\.html$/, ''),
+                );
+            });
+
+            return async.map(
+                partials,
+                templateAssembler.assembleAndBundle.bind(null, this.templatesPath),
+                (err, results) => {
+                    const ret = {};
+
+                    if (err) {
+                        return callback(err);
+                    }
+
+                    partials.forEach((file, index) => {
+                        ret[file] = results[index];
+                    });
+
+                    return async.parallel(
+                        [
+                            this._checkObjects.bind(this, results),
+                            this._detectCycles.bind(this, results),
+                        ],
+                        (checkingError) => {
+                            if (checkingError) {
+                                callback(checkingError);
+                            }
+
+                            console.log(`${'ok'.green} -- Template Parsing Finished`);
+                            callback(null, ret);
+                        },
+                    );
+                },
+            );
+        });
+    }
+
+    async assembleSchema() {
+        console.log('Building Theme Schema File...');
+
+        const schema = await this.themeConfig.getSchema();
+
+        console.log(`${'ok'.green} -- Theme Schema Building Finished`);
+
+        return schema;
+    }
+
+    async assembleSchemaTranslations() {
+        console.log('Schema Translations Parsing Started...');
+
+        const schema = await this.themeConfig.getSchemaTranslations();
+
+        console.log(`${'ok'.green} -- Schema Translations Parsing Finished`);
+
+        return schema;
+    }
+
+    assembleLangTask(callback) {
+        console.log('Language Files Parsing Started...');
+        langAssembler.assemble((err, results) => {
             if (err) {
                 return callback(err);
             }
 
-            taskResults.manifest = manifest;
-            // zip theme files
-            bundleThemeFiles(archive, this.themePath, this.configuration);
+            console.log(`${'ok'.green} -- Language Files Parsing Finished`);
+            return callback(null, results);
+        });
+    }
 
-            // zip all generated files
-            const failedTemplates = bundleParsedFiles(archive, taskResults);
+    generateManifest(taskResults, callback) {
+        console.log('Generating Manifest Started...');
 
-            fileStream.on('close', () => {
-                const stats = Fs.statSync(bundleZipPath);
-                const size = stats['size'];
+        recursiveReadDir(this.templatesPath, ['!*.html'], (err, filePaths) => {
+            if (err) {
+                return callback(err);
+            }
 
-                if (failedTemplates.length) {
-                    return console.error(`Error: Your bundle failed as templates generated from the files below are greater than or equal to 1 megabyte in size:\n${failedTemplates.join('\n')}`);
-                }
-
-                if (size > MAX_SIZE_BUNDLE) {
-                    return console.error(`Error: Your bundle of size ${size} bytes is above the max size of ${MAX_SIZE_BUNDLE} bytes`);
-                }
-
-                console.log('ok'.green + ' -- Zipping Files Finished');
-
-                return callback(null, bundleZipPath);
+            const templates = filePaths.map((file) => {
+                return upath.toUnix(
+                    file.replace(this.templatesPath + path.sep, '').replace('.html', ''),
+                );
             });
 
-            // This triggers 'close' event in the file stream. No need to callback()
-            archive.finalize();
-        });
-    });
-}
+            const regions = fetchRegions(taskResults.templates, templates);
 
-/**
- * Archive theme files
- * @param {Archiver} archive
- * @param {String} themePath
- * @param {Object} configuration
- */
-function bundleThemeFiles(archive, themePath, configuration) {
-    if (configuration.jspm) {
-        archive.append(
-            Fs.createReadStream(configuration.jspm.tmpBundleFile),
-            { name: configuration.jspm.bundle_location },
-        );
+            console.log(`${'ok'.green} -- Manifest Generation Finished`);
+            return callback(null, {
+                regions,
+                templates,
+            });
+        });
     }
 
-    archive.bulk({
-        src: PATHS_TO_ZIP,
-        cwd: themePath,
-        expand: true,
-    });
-}
-
-/**
- * Archive all generated files (ex. parsed files)
- * @param {Archiver} archive
- * @param {Object} taskResults
- * @returns {Array}
- */
-function bundleParsedFiles(archive, taskResults) {
-    const archiveJsonFile = (data, name) => {
-        archive.append(JSON.stringify(data, null, 2), { name });
-    };
-    const failedTemplates = [];
-    for (let task in taskResults) {
-        let data = taskResults[task];
-        switch(task) {
-        case 'css':
-            // Create the parsed tree files
-            for (let filename in data) {
-                archiveJsonFile(data[filename], `parsed/scss/${filename}.json`);
+    /**
+     * @private
+     * @param {Object[]} results
+     * @param {Function} callback
+     * @returns {void}
+     */
+    _checkObjects(results, callback) {
+        this.validator.validateObjects(results, (err) => {
+            if (err) {
+                console.error('error '.red + err.message);
+                return callback(err);
             }
-            break;
 
-        case 'templates':
-            // Create the parsed tree files
-            for (let filename in data) {
-                const hash = Crypto.createHash('md5').update(filename).digest('hex');
-                const fileData = data[filename];
-                archiveJsonFile(fileData, `parsed/templates/${hash}.json`);
-                // if file size is greater than 1 megabyte push filename to failedTemplates
-                if (JSON.stringify(fileData).length >= MEGABYTE) {
-                    failedTemplates.push(filename);
-                }
-            }
-            break;
+            return callback();
+        });
+    }
 
-        case 'lang':
-            // append the parsed translation file with all translations
-            archiveJsonFile(data, 'parsed/lang.json');
-            break;
-
-        case 'schema':
-            // append the generated schema.json file
-            archiveJsonFile(data, 'schema.json');
-            break;
-
-        case 'schemaTranslations':
-            // append the parsed schemaTranslations.json file
-            archiveJsonFile(data, 'schemaTranslations.json');
-            break;
-
-        case 'manifest':
-            // append the generated manifest.json file
-            archiveJsonFile(data, 'manifest.json');
-            break;
+    /**
+     * @private
+     * @param {Object[]} results
+     * @param {Function} callback
+     * @returns {void}
+     */
+    _detectCycles(results, callback) {
+        try {
+            new Cycles(results).detect();
+            callback();
+        } catch (err) {
+            callback(err);
         }
     }
-    return failedTemplates;
+
+    /**
+     * @private
+     * @param {Function} callback
+     * @returns {void}
+     */
+    _validateTheme(callback) {
+        console.log('Validating theme...');
+
+        this.validator.validateTheme((err) => {
+            if (err) {
+                throw err;
+            }
+            callback(null, true);
+        });
+    }
+
+    /**
+     * @private
+     * @param {Function} callback
+     * @returns {void}
+     */
+    _bundleTaskRunner(callback) {
+        const defaultName = this.configuration.name
+            ? `${this.configuration.name}-${this.configuration.version}.zip`
+            : 'Theme.zip';
+        const outputName =
+            typeof this.options.name === 'string' ? `${this.options.name}.zip` : defaultName;
+        const outputFolder =
+            typeof this.options.dest === 'string' ? this.options.dest : this.themePath;
+        const bundleZipPath = path.join(outputFolder, outputName);
+
+        async.parallel(this.tasks, (err, taskResults) => {
+            if (err) {
+                return callback(err);
+            }
+
+            const archive = Archiver('zip');
+            const fileStream = fs.createWriteStream(bundleZipPath);
+            archive.pipe(fileStream);
+
+            // Create manifest will use taskResults to generate a manifest file
+            return this.generateManifest(taskResults, (manifestGenerationError, manifest) => {
+                if (manifestGenerationError) {
+                    return callback(manifestGenerationError);
+                }
+
+                // eslint-disable-next-line no-param-reassign
+                taskResults.manifest = manifest;
+                // zip theme files
+                this._bundleThemeFiles(archive, this.themePath);
+
+                // zip all generated files
+                const failedTemplates = this._bundleParsedFiles(archive, taskResults);
+
+                fileStream.on('close', () => {
+                    const stats = fs.statSync(bundleZipPath);
+                    const { size } = stats;
+
+                    if (failedTemplates.length) {
+                        return console.error(
+                            `Error: Your bundle failed as templates generated from the files below are greater than or equal to 1 megabyte in size:\n${failedTemplates.join(
+                                '\n',
+                            )}`,
+                        );
+                    }
+
+                    if (size > MAX_SIZE_BUNDLE) {
+                        return console.error(
+                            `Error: Your bundle of size ${size} bytes is above the max size of ${MAX_SIZE_BUNDLE} bytes`,
+                        );
+                    }
+
+                    console.log(`${'ok'.green} -- Zipping Files Finished`);
+
+                    return callback(null, bundleZipPath);
+                });
+
+                // This triggers 'close' event in the file stream. No need to callback()
+                return archive.finalize();
+            });
+        });
+    }
+
+    /**
+     * Archive theme files
+     * @private
+     * @param {Archiver} archive
+     * @param {string} themePath
+     * @returns {void}
+     */
+    _bundleThemeFiles(archive, themePath) {
+        for (const { pattern, ignore } of PATHS_TO_ZIP) {
+            archive.glob(pattern, { ignore, cwd: themePath });
+        }
+    }
+
+    /**
+     * @private
+     * Archive all generated files (ex. parsed files)
+     * @param {Archiver} archive
+     * @param {Object} taskResults
+     * @returns {Array}
+     */
+    _bundleParsedFiles(archive, taskResults) {
+        const archiveJsonFile = (data, name) => {
+            archive.append(JSON.stringify(data, null, 2), { name });
+        };
+        const failedTemplates = [];
+
+        for (const [task, result] of Object.entries(taskResults)) {
+            switch (task) {
+                case 'css':
+                    // Create the parsed tree files
+                    for (const [filename, data] of Object.entries(result)) {
+                        archiveJsonFile(data, `parsed/scss/${filename}.json`);
+                    }
+                    break;
+
+                case 'templates':
+                    // Create the parsed tree files
+                    for (const [filename, data] of Object.entries(result)) {
+                        const hash = crypto.createHash('md5').update(filename).digest('hex');
+                        const fileData = data;
+                        archiveJsonFile(fileData, `parsed/templates/${hash}.json`);
+                        // if file size is greater than 1 megabyte push filename to failedTemplates
+                        if (JSON.stringify(fileData).length >= MEGABYTE) {
+                            failedTemplates.push(filename);
+                        }
+                    }
+                    break;
+
+                case 'lang':
+                    // append the parsed translation file with all translations
+                    archiveJsonFile(result, 'parsed/lang.json');
+                    break;
+
+                case 'schema':
+                    // append the generated schema.json file
+                    archiveJsonFile(result, 'schema.json');
+                    break;
+
+                case 'schemaTranslations':
+                    // append the parsed schemaTranslations.json file
+                    archiveJsonFile(result, 'schemaTranslations.json');
+                    break;
+
+                case 'manifest':
+                    // append the generated manifest.json file
+                    archiveJsonFile(result, 'manifest.json');
+                    break;
+
+                default:
+                    break;
+            }
+        }
+        return failedTemplates;
+    }
 }
 
 module.exports = Bundle;
