@@ -1,118 +1,152 @@
-'use strict';
-var Fs = require('fs');
-var jsonLint = require('./json-lint');
-var Path = require('path');
-var Inquirer = require('inquirer');
-var themePath = process.cwd();
-var hoek = require('hoek');
-var configuration;
-var internals = {};
+require('colors');
 
-internals.parseAnswers = function(JspmAssembler, ThemeConfig, dotStencilFile, dotStencilFilePath, answers) {
+const inquirerModule = require('inquirer');
 
-    // Check for custom layout configurations
-    // If already set, do nothing otherwise write the empty configurations
-    if (!dotStencilFile || dotStencilFile && !dotStencilFile.customLayouts) {
-        answers.customLayouts = {
-            'brand': {},
-            'category': {},
-            'page': {},
-            'product': {},
+const serverConfigModule = require('../server/config');
+const StencilConfigManager = require('./StencilConfigManager');
+const { DEFAULT_CUSTOM_LAYOUTS_CONFIG } = require('../constants');
+
+class StencilInit {
+    /**
+     * @param inquirer
+     * @param stencilConfigManager
+     * @param serverConfig
+     * @param logger
+     */
+    constructor({
+        inquirer = inquirerModule,
+        stencilConfigManager = new StencilConfigManager(),
+        serverConfig = serverConfigModule,
+        logger = console,
+    } = {}) {
+        this._inquirer = inquirer;
+        this._stencilConfigManager = stencilConfigManager;
+        this._serverConfig = serverConfig;
+        this._logger = logger;
+    }
+
+    /**
+     * @param {object} cliOptions
+     * @param {string} cliOptions.normalStoreUrl
+     * @param {string} cliOptions.accessToken
+     * @param {number} cliOptions.port
+     * @returns {Promise<void>}
+     */
+    async run(cliOptions = {}) {
+        const oldStencilConfig = await this.readStencilConfig();
+        const defaultAnswers = this.getDefaultAnswers(oldStencilConfig);
+        const questions = this.getQuestions(defaultAnswers, cliOptions);
+        const answers = await this.askQuestions(questions);
+        const updatedStencilConfig = this.applyAnswers(oldStencilConfig, answers, cliOptions);
+        await this._stencilConfigManager.save(updatedStencilConfig);
+
+        this._logger.log(
+            'You are now ready to go! To start developing, run $ ' + 'stencil start'.cyan,
+        );
+    }
+
+    /**
+     * @returns {object}
+     */
+    async readStencilConfig() {
+        let parsedConfig;
+
+        try {
+            parsedConfig = await this._stencilConfigManager.read(true, true);
+        } catch (err) {
+            this._logger.error(
+                'Detected a broken stencil-cli config:\n',
+                err,
+                '\nThe file will be rewritten with your answers',
+            );
+        }
+
+        return parsedConfig || {};
+    }
+
+    /**
+     * @param {{port: (number), normalStoreUrl: (string), accessToken: (string)}} stencilConfig
+     * @returns {{port: (number), normalStoreUrl: (string), accessToken: (string)}}
+     */
+    getDefaultAnswers(stencilConfig) {
+        return {
+            normalStoreUrl: stencilConfig.normalStoreUrl,
+            accessToken: stencilConfig.accessToken,
+            port: stencilConfig.port || this._serverConfig.get('/server/port'),
         };
     }
 
-    var defaults = dotStencilFile ? hoek.applyToDefaults(dotStencilFile, answers) : answers;
+    /**
+     * @param {{port: (number), normalStoreUrl: (string), accessToken: (string)}} defaultAnswers
+     * @param {{port: (number), normalStoreUrl: (string), accessToken: (string)}} cliOptions
+     * @returns {{object[]}}
+     */
+    getQuestions(defaultAnswers, cliOptions) {
+        const prompts = [];
 
-    Fs.writeFile(dotStencilFilePath, JSON.stringify(defaults, null, 2), function (err) {
-        var ready = 'You are now ready to go! To start developing, run $ ' + 'stencil start'.cyan,
-            bundleTask;
-
-        if (err) {
-            throw err;
-        }
-
-        // bundle dev dependencies
-        configuration = ThemeConfig.getInstance(themePath).getConfig();
-        if (configuration.jspm) {
-            if (!Fs.existsSync(Path.join(themePath, configuration.jspm.jspm_packages_path))) {
-                console.log('Error: The path you specified for your "jspm_packages" folder does not exist.'.red);
-                return console.log(
-                    'Please check your '.red +
-                    'jspm.jspm_packages_path'.cyan +
-                    ' setting in your theme\'s '.red +
-                    'config.json'.cyan +
-                    ' file to make sure it\'s correct.'.red,
-                );
-            }
-
-            bundleTask = JspmAssembler.assemble(configuration.jspm, themePath);
-
-            bundleTask(function () {
-                console.log(ready);
+        if (!cliOptions.normalStoreUrl) {
+            prompts.push({
+                type: 'input',
+                name: 'normalStoreUrl',
+                message: "What is the URL of your store's home page?",
+                validate: (val) => /^https?:\/\//.test(val) || 'You must enter a URL',
+                default: defaultAnswers.normalStoreUrl,
             });
-
-        } else {
-            console.log(ready);
         }
-    });
-};
 
-internals.implementation = function(JspmAssembler, ThemeConfig, dotStencilFilePath, url, token, port) {
-    var dotStencilFile;
-    var questions;
-
-    if (Fs.existsSync(dotStencilFilePath)) {
-        dotStencilFile = Fs.readFileSync(dotStencilFilePath, {encoding: 'utf-8'});
-        try {
-            dotStencilFile = jsonLint.parse(dotStencilFile, dotStencilFilePath);
-        } catch (e) {
-            return console.error(e.fileName, e.stack);
+        if (!cliOptions.accessToken) {
+            prompts.push({
+                type: 'input',
+                name: 'accessToken',
+                message: 'What is your Stencil OAuth Access Token?',
+                default: defaultAnswers.accessToken,
+                filter: (val) => val.trim(),
+            });
         }
+
+        if (!cliOptions.port) {
+            prompts.push({
+                type: 'input',
+                name: 'port',
+                message: 'What port would you like to run the server on?',
+                default: defaultAnswers.port,
+                validate: (val) => {
+                    if (Number.isNaN(val)) {
+                        return 'You must enter an integer';
+                    }
+                    if (val < 1024 || val > 65535) {
+                        return 'The port number must be between 1025 and 65535';
+                    }
+                    return true;
+                },
+            });
+        }
+
+        return prompts;
     }
 
-    questions = [
-        {
-            type: 'input',
-            name: 'normalStoreUrl',
-            message: 'What is the URL of your store\'s home page?',
-            validate: function (val) {
-                if (/^https?:\/\//.test(val)) {
-                    return true;
-                } else {
-                    return 'You must enter a URL';
-                }
-            },
-            default: url || dotStencilFile && dotStencilFile.normalStoreUrl || undefined,
-        },
-        {
-            type: 'input',
-            name: 'accessToken',
-            message: 'What is your Stencil OAuth Access Token?',
-            default: token || dotStencilFile && dotStencilFile.accessToken,
-            filter: function(val) {
-                return val.trim();
-            },
-        },
-        {
-            type: 'input',
-            name: 'port',
-            message: 'What port would you like to run the server on?',
-            default: port || dotStencilFile && dotStencilFile.port || 3000,
-            validate: function (val) {
-                if (isNaN(val)) {
-                    return 'You must enter an integer';
-                } else if (val < 1024 || val > 65535) {
-                    return 'The port number must be between 1025 and 65535';
-                } else {
-                    return true;
-                }
-            },
-        },
-    ];
+    /**
+     * @param {{object[]}} questions
+     * @returns {Promise<object>}
+     */
+    async askQuestions(questions) {
+        return questions.length ? this._inquirer.prompt(questions) : {};
+    }
 
-    Inquirer.prompt(questions, function(answers) {
-        internals.parseAnswers(JspmAssembler, ThemeConfig, dotStencilFile, dotStencilFilePath, answers);
-    });
-};
+    /**
+     * @param {object} stencilConfig
+     * @param {object} answers
+     * @param {object} cliOptions
+     * @returns {object}
+     */
+    applyAnswers(stencilConfig, answers, cliOptions) {
+        return {
+            customLayouts: DEFAULT_CUSTOM_LAYOUTS_CONFIG,
+            ...stencilConfig,
+            ...cliOptions,
+            ...answers,
+        };
+    }
+}
 
-module.exports = internals.implementation;
+module.exports = StencilInit;
