@@ -1,20 +1,21 @@
-'use strict';
 const async = require('async');
-let stencilPushUtils = require('./stencil-push.utils');
-let stencilPullUtils = require('./stencil-pull.utils');
+const stencilPushUtils = require('./stencil-push.utils');
+const stencilPullUtils = require('./stencil-pull.utils');
+
+function stencilPull(options = {}, callback) {
+    async.waterfall(
+        [
+            async.constant(options),
+            stencilPushUtils.readStencilConfigFile,
+            stencilPushUtils.getStoreHash,
+            stencilPushUtils.getChannels,
+            stencilPushUtils.promptUserForChannel,
+            stencilPullUtils.getChannelActiveTheme,
+            stencilPullUtils.getThemeConfiguration,
+            stencilPullUtils.mergeThemeConfiguration,
+        ],
+        callback,
+    );
+}
 
 module.exports = stencilPull;
-
-function stencilPull(options, callback) {
-    options = options || {};
-    async.waterfall([
-        async.constant(options),
-        stencilPushUtils.readStencilConfigFile,
-        stencilPushUtils.getStoreHash,
-        stencilPushUtils.getThemes,
-        stencilPullUtils.selectActiveTheme,
-        stencilPullUtils.startThemeDownloadJob,
-        stencilPushUtils.pollForJobCompletion(({ download_url: downloadUrl }) => ({ downloadUrl })),
-        stencilPullUtils.downloadThemeConfig,
-    ], callback);
-}
