@@ -1,106 +1,107 @@
-const themeApiClient = require('./theme-api-client');
-const request = require("request");
-const yauzl = require('yauzl');
 const fs = require('fs');
-const path = require('path');
-const tmp = require('tmp');
+const _ = require('lodash');
+const themeApiClient = require('./theme-api-client');
+const { parseJsonFile } = require('./utils/fsUtils');
+
 const utils = {};
 
+utils.getChannelActiveTheme = async (options) => {
+    const {
+        config: { accessToken },
+        apiHost,
+        storeHash,
+        channelId,
+    } = options;
+
+    const activeTheme = await themeApiClient.getChannelActiveTheme({
+        accessToken,
+        apiHost,
+        storeHash,
+        channelId,
+    });
+
+    console.log('ok'.green + ` -- Fetched theme details for channel ${channelId}`);
+
+    return { ...options, activeTheme };
+};
+
+utils.getThemeConfiguration = async (options) => {
+    const {
+        config: { accessToken },
+        apiHost,
+        storeHash,
+        activeTheme,
+        saved,
+    } = options;
+
+    const themeId = activeTheme.active_theme_uuid;
+
+    const configurationId = saved
+        ? activeTheme.saved_theme_configuration_uuid
+        : activeTheme.active_theme_configuration_uuid;
+
+    const remoteThemeConfiguration = await themeApiClient.getThemeConfiguration({
+        accessToken,
+        apiHost,
+        storeHash,
+        themeId,
+        configurationId,
+    });
+
+    console.log('ok'.green + ` -- Fetched ${saved ? 'saved' : 'active'} configuration`);
+
+    return { ...options, remoteThemeConfiguration };
+};
+
+utils.mergeThemeConfiguration = async (options) => {
+    const { remoteThemeConfiguration } = options;
+
+    const localConfig = await parseJsonFile('config.json');
+    let diffDetected = false;
+
+    // For any keys the remote configuration has in common with the local configuration,
+    // overwrite the local configuration if the remote configuration differs
+    for (const [key, remoteVal] of Object.entries(remoteThemeConfiguration.settings)) {
+        if (!(key in localConfig.settings)) {
+            continue;
+        }
+        const localVal = localConfig.settings[key];
+
+        // Check for different types, and throw an error if they are found
+        if (typeof localVal !== typeof remoteVal) {
+            throw new Error(
+                `Theme configuration key "${key}" cannot be merged because it is not of the same type. ` +
+                    `Remote configuration is of type ${typeof remoteVal} while local configuration is of type ${typeof localVal}.`,
+            );
+        }
+
+        // If a different value is found, overwrite the local config
+        if (!_.isEqual(localVal, remoteVal)) {
+            localConfig.settings[key] = remoteVal;
+            diffDetected = true;
+        }
+    }
+
+    // Does a file need to be written?
+    if (diffDetected || options.saveConfigName !== 'config.json') {
+        if (diffDetected) {
+            console.log('ok'.green + ' -- Remote configuration merged with local configuration');
+        } else {
+            console.log(
+                'ok'.green + ' -- Remote and local configurations are in sync for all common keys',
+            );
+        }
+
+        await fs.promises.writeFile(options.saveConfigName, JSON.stringify(localConfig, null, 2));
+        console.log('ok'.green + ` -- Configuration written to ${options.saveConfigName}`);
+    } else {
+        console.log(
+            'ok'.green +
+                ` -- Remote and local configurations are in sync for all common keys, no action taken`,
+        );
+    }
+
+    return options;
+};
+
 module.exports = utils;
-
-utils.selectActiveTheme = (options, callback) => {
-    const [activeTheme] = options.themes.filter(theme => theme.is_active).map(theme => theme.uuid);
-
-    callback(null, Object.assign({}, options, { activeTheme }));
-};
-
-utils.startThemeDownloadJob = (options, callback) => {
-    const config = options.config;
-
-    themeApiClient.downloadTheme({
-        accessToken: config.accessToken,
-        apiHost: options.apiHost,
-        themeId: options.activeTheme,
-        clientId: 'stencil-cli',
-        storeHash: options.storeHash,
-    }, (error, result) => {
-        if (error) {
-            error.name = 'ThemeUploadError';
-            return callback(error);
-        }
-
-        callback(null, Object.assign({}, options, {
-            jobId: result.jobId,
-        }));
-    });
-};
-
-utils.downloadThemeConfig = (options, callback) => {
-    tmp.file(function _tempFileCreated(err, tempThemePath, fd, cleanupCallback) {
-        if (err) {
-            callback(err);
-        }
-
-        (
-            new Promise(
-                (resolve, reject) =>
-                    request(options.downloadUrl)
-                        .pipe(fs.createWriteStream(tempThemePath))
-                        .on('finish', () => resolve(tempThemePath))
-                        .on('error', reject),
-            )
-        )
-            .then(tempThemePath =>
-                new Promise(
-                    (resolve, reject) =>
-                        yauzl.open(tempThemePath, { lazyEntries: true }, (error, zipFile) => {
-                            if (error) {
-                                return reject(error);
-                            }
-
-                            zipFile.readEntry();
-                            zipFile.on('entry', entry => {
-                                if (!/config\.json/.test(entry.fileName)) {
-                                    zipFile.readEntry();
-                                    return;
-                                }
-
-                                zipFile.openReadStream(entry, (readStreamError, readStream) => {
-                                    if (readStreamError) {
-                                        return reject(readStreamError);
-                                    }
-
-                                    let configFileData = '';
-
-                                    readStream.on('end', () => {
-                                        resolve(JSON.parse(configFileData));
-                                        zipFile.close();
-                                    });
-                                    readStream.on('data', chunk => {
-                                        configFileData += chunk;
-                                    });
-                                });
-                            });
-                        }),
-                ),
-            )
-            .then(
-                liveStencilConfig =>
-                    new Promise(
-                        (resolve, reject) =>
-                            fs.writeFile(path.resolve(options.saveConfigName), JSON.stringify(liveStencilConfig, null, 2), error => {
-                                if (error) {
-                                    reject(error);
-                                }
-
-                                resolve();
-                            }),
-                    ),
-            )
-            .then(() => {
-                cleanupCallback();
-                callback(null, options);
-            })
-            .catch(callback);
-    });
-};
