@@ -1,23 +1,24 @@
-'use strict';
 const async = require('async');
-let utils = require('./stencil-push.utils');
+const utils = require('./stencil-push.utils');
+
+function stencilPush(options = {}, callback) {
+    async.waterfall(
+        [
+            async.constant(options),
+            utils.readStencilConfigFile,
+            utils.getStoreHash,
+            utils.getThemes,
+            utils.generateBundle,
+            utils.uploadBundle,
+            utils.notifyUserOfThemeLimitReachedIfNecessary,
+            utils.promptUserToDeleteThemesIfNecessary,
+            utils.deleteThemesIfNecessary,
+            utils.uploadBundleAgainIfNecessary,
+            utils.notifyUserOfThemeUploadCompletion,
+            utils.pollForJobCompletion((data) => ({ themeId: data.theme_id })),
+        ],
+        callback,
+    );
+}
 
 module.exports = stencilPush;
-
-function stencilPush(options, callback) {
-    options = options || {};
-    async.waterfall([
-        async.constant(options),
-        utils.readStencilConfigFile,
-        utils.getStoreHash,
-        utils.getThemes,
-        utils.generateBundle,
-        utils.uploadBundle,
-        utils.notifyUserOfThemeLimitReachedIfNecessary,
-        utils.promptUserToDeleteThemesIfNecessary,
-        utils.deleteThemesIfNecessary,
-        utils.uploadBundleAgainIfNecessary,
-        utils.notifyUserOfThemeUploadCompletion,
-        utils.pollForJobCompletion(data => ({ themeId: data.theme_id })),
-    ], callback);
-}
