@@ -1,17 +1,19 @@
-'use strict';
 const _ = require('lodash');
 const async = require('async');
-const Bundle = require('./stencil-bundle');
-const fs = require('fs');
 const Inquirer = require('inquirer');
-const os = require('os');
 const ProgressBar = require('progress');
-const themeApiClient = require('./theme-api-client');
-const themePath = process.cwd();
-const themeConfig = require('./theme-config').getInstance(themePath);
 const uuid = require('uuid4');
+const os = require('os');
+
+const { THEME_PATH } = require('../constants');
+const Bundle = require('./stencil-bundle');
+const themeApiClient = require('./theme-api-client');
+const ThemeConfig = require('./theme-config');
+const StencilConfigManager = require('./StencilConfigManager');
+
+const themeConfigManager = ThemeConfig.getInstance(THEME_PATH);
+const stencilConfigManager = new StencilConfigManager();
 const utils = {};
-const Wreck = require('wreck');
 
 const bar = new ProgressBar('Processing [:bar] :percent; ETA: :etas', {
     complete: '=',
@@ -19,362 +21,414 @@ const bar = new ProgressBar('Processing [:bar] :percent; ETA: :etas', {
     total: 100,
 });
 
-module.exports = utils;
-
-function validateOptions(options, fields) {
-    options = options || {};
-    fields = fields || [];
-
-    fields.forEach(field => {
+function validateOptions(options = {}, fields = []) {
+    for (const field of fields) {
         if (!_.has(options, field)) {
             throw new Error(`${field} is required!`);
         }
-    });
-
-    return options;
+    }
 }
 
-utils.readStencilConfigFile = (options, callback) => {
-    options = validateOptions(options, ['dotStencilFilePath']);
-
-    fs.readFile(options.dotStencilFilePath, { encoding: 'utf8' }, (err, data) => {
-        if (err) {
-            err.name = 'StencilConfigReadError';
-            return callback(err);
-        }
-
-        callback(null, Object.assign({}, options, {
-            config: JSON.parse(data),
-        }));
-    });
+utils.readStencilConfigFile = async (options) => {
+    try {
+        const config = await stencilConfigManager.read();
+        return { ...options, config };
+    } catch (err) {
+        err.name = 'StencilConfigReadError';
+        throw err;
+    }
 };
 
-utils.getStoreHash = (options, callback) => {
-    options = validateOptions(options, ['config.normalStoreUrl']);
+utils.getStoreHash = async (options) => {
+    validateOptions(options, ['config.normalStoreUrl']);
 
-    Wreck.get(`https://${options.config.normalStoreUrl.replace(/http(s?):\/\//, '').split('/')[0]}/admin/oauth/info`, {
-        json: true,
-        rejectUnauthorized: false,
-    }, (error, response, payload) => {
-        if (error) {
-            error.name = 'StoreHashReadError';
-            return callback(error);
-        }
-
-        if (response.statusCode !== 200 || !payload.store_hash) {
-            const err = new Error('Failed to retrieve store hash');
-            err.name = 'StoreHashReadError';
-            return callback(err);
-        }
-
-        callback(null, Object.assign({}, options, { storeHash: payload.store_hash }));
+    const storeHash = await themeApiClient.getStoreHash({
+        storeUrl: options.config.normalStoreUrl,
     });
+
+    return { ...options, storeHash };
 };
 
-utils.getThemes = (options, callback) => {
-    const config = options.config;
+utils.getThemes = async (options) => {
+    const {
+        config: { accessToken },
+        apiHost,
+        storeHash,
+    } = options;
 
-    themeApiClient.getThemes({
-        accessToken: config.accessToken,
-        apiHost: options.apiHost,
-        clientId: 'stencil-cli',
-        storeHash: options.storeHash,
-    }, (error, result) => {
-        if (error) {
-            return callback(error);
-        }
+    const themes = await themeApiClient.getThemes({ accessToken, apiHost, storeHash });
 
-        callback(null, Object.assign({}, options, {
-            themes: result.themes,
-        }));
-    });
+    return { ...options, themes };
 };
 
-utils.generateBundle = (options, callback) => {
-    let bundle;
-    let output = {
-        dest: os.tmpdir(),
-        name: uuid(),
-    };
-
+utils.generateBundle = async (options) => {
     if (options.bundleZipPath) {
-        return async.nextTick(callback.bind(null, null, options));
+        return options;
     }
 
-    if (options.saveBundleName) {
-        output = {
-            dest: themePath,
-            name: options.saveBundleName,
-        };
+    const output = options.saveBundleName
+        ? { dest: THEME_PATH, name: options.saveBundleName }
+        : { dest: os.tmpdir(), name: uuid() };
+    const rawConfig = await themeConfigManager.getRawConfig();
+    const bundle = new Bundle(THEME_PATH, themeConfigManager, rawConfig, output);
+
+    try {
+        const bundleZipPath = await bundle.initBundle();
+        return { ...options, bundleZipPath };
+    } catch (err) {
+        err.name = 'BundleInitError';
+        throw err;
     }
-
-    bundle = new Bundle(themePath, themeConfig, themeConfig.getRawConfig(), output);
-
-    bundle.initBundle((err, bundleZipPath) => {
-        if (err) {
-            err.name = 'BundleInitError';
-            return callback(err);
-        }
-
-        callback(null, Object.assign(options, { bundleZipPath: options.bundleZipPath || bundleZipPath }));
-    });
 };
 
-utils.uploadBundle = (options, callback) => {
-    const config = options.config;
+utils.uploadBundle = async (options) => {
+    const {
+        config: { accessToken },
+        apiHost,
+        bundleZipPath,
+        storeHash,
+    } = options;
 
     console.log('info'.blue + ` -- Uploading bundle located here: ${options.bundleZipPath}`);
 
-    themeApiClient.postTheme({
-        accessToken: config.accessToken,
-        apiHost: options.apiHost,
-        bundleZipPath: options.bundleZipPath,
-        clientId: 'stencil-cli',
-        storeHash: options.storeHash,
-    }, (error, result) => {
-        if (error) {
-            error.name = 'ThemeUploadError';
-            return callback(error);
-        }
-        callback(null, Object.assign({}, options, {
+    try {
+        const result = await themeApiClient.postTheme({
+            accessToken,
+            apiHost,
+            bundleZipPath,
+            storeHash,
+        });
+        return {
+            ...options,
             jobId: result.jobId,
             themeLimitReached: !!result.themeLimitReached,
-        }));
-    });
+        };
+    } catch (err) {
+        err.name = 'ThemeUploadError';
+        throw err;
+    }
 };
 
-utils.notifyUserOfThemeLimitReachedIfNecessary = (options, callback) => {
+utils.notifyUserOfThemeLimitReachedIfNecessary = async (options) => {
     if (options.themeLimitReached && !options.deleteOldest) {
-        console.log('warning'.yellow + ` -- You have reached your upload limit.  In order to proceed, you'll need to delete at least one theme.`);
+        console.log(
+            'warning'.yellow +
+                ' -- You have reached your upload limit. ' +
+                "In order to proceed, you'll need to delete at least one theme.",
+        );
     }
 
-    return async.nextTick(callback.bind(null, null, options));
+    return options;
 };
 
-utils.promptUserToDeleteThemesIfNecessary = (options, callback) => {
+utils.promptUserToDeleteThemesIfNecessary = async (options) => {
     if (!options.themeLimitReached) {
-        return async.nextTick(callback.bind(null, null, options));
+        return options;
     }
 
     if (options.deleteOldest) {
         const oldestTheme = options.themes
-            .filter(theme => theme.is_private && !theme.is_active)
-            .map(theme => ({ uuid: theme.uuid, updated_at: new Date(theme.updated_at).valueOf() }))
-            .reduce((prev, current) => prev.updated_at < current.updated_at ? prev : current);
-        return callback(null, Object.assign({}, options, { themeIdsToDelete: [oldestTheme.uuid] }));
+            .filter((theme) => theme.is_private && !theme.is_active)
+            .map((theme) => ({
+                uuid: theme.uuid,
+                updated_at: new Date(theme.updated_at).valueOf(),
+            }))
+            .reduce((prev, current) => (prev.updated_at < current.updated_at ? prev : current));
+
+        return { ...options, themeIdsToDelete: [oldestTheme.uuid] };
     }
 
-    const questions = [{
-        choices: options.themes.map(theme => ({
-            disabled: theme.is_active || !theme.is_private,
-            name: theme.name,
-            value: theme.uuid,
-        })),
-        message: 'Which theme(s) would you like to delete?',
-        name: 'themeIdsToDelete',
-        type: 'checkbox',
-        validate: val => {
-            if (val.length > 0) {
-                return true;
-            } else {
+    const questions = [
+        {
+            choices: options.themes.map((theme) => ({
+                disabled: theme.is_active || !theme.is_private,
+                name: theme.name,
+                value: theme.uuid,
+            })),
+            message: 'Which theme(s) would you like to delete?',
+            name: 'themeIdsToDelete',
+            type: 'checkbox',
+            validate: (val) => {
+                if (val.length > 0) {
+                    return true;
+                }
                 return 'You must delete at least one theme';
-            }
+            },
         },
-    }];
+    ];
+    const answers = await Inquirer.prompt(questions);
 
-    Inquirer.prompt(questions, answers => {
-        callback(null, Object.assign({}, options, answers));
-    });
+    return { ...options, ...answers };
 };
 
-utils.deleteThemesIfNecessary = (options, callback) => {
-    const config = options.config;
+utils.deleteThemesIfNecessary = async (options) => {
+    const {
+        config: { accessToken },
+        apiHost,
+        storeHash,
+        themeLimitReached,
+        themeIdsToDelete,
+    } = options;
 
-    if (!options.themeLimitReached) {
-        return async.nextTick(callback.bind(null, null, options));
+    if (!themeLimitReached) {
+        return options;
     }
 
-    async.parallel(options.themeIdsToDelete.map(themeId => {
-        return cb => {
-            themeApiClient.deleteThemeById(Object.assign({
-                accessToken: config.accessToken,
-                apiHost: options.apiHost,
-                clientId: 'stencil-cli',
-                storeHash: options.storeHash,
-                themeId,
-            }, options), cb);
-        };
-    }), err => {
-        if (err) {
-            err.name = 'ThemeDeletionError';
-            return callback(err);
-        }
-
-        callback(null, options);
-    });
-};
-
-utils.uploadBundleAgainIfNecessary = (options, callback) => {
-    if (!options.themeLimitReached) {
-        return async.nextTick(callback.bind(null, null, options));
+    try {
+        const promises = themeIdsToDelete.map((themeId) =>
+            themeApiClient.deleteThemeById({ accessToken, apiHost, storeHash, themeId }),
+        );
+        await Promise.all(promises);
+    } catch (err) {
+        err.name = 'ThemeDeletionError';
+        throw err;
     }
 
-    utils.uploadBundle(options, callback);
+    return options;
 };
 
-utils.notifyUserOfThemeUploadCompletion = (options, callback) => {
-    console.log('ok'.green + ' -- Theme Upload Finished');
-    return async.nextTick(callback.bind(null, null, options));
+utils.uploadBundleAgainIfNecessary = async (options) => {
+    if (!options.themeLimitReached) {
+        return options;
+    }
+
+    return utils.uploadBundle(options);
 };
 
-utils.markJobProgressPercentage = percentComplete => {
+utils.notifyUserOfThemeUploadCompletion = async (options) => {
+    console.log(`${'ok'.green} -- Theme Upload Finished`);
+
+    return options;
+};
+
+utils.markJobProgressPercentage = (percentComplete) => {
     bar.update(percentComplete / 100);
 };
 
 utils.markJobComplete = () => {
     utils.markJobProgressPercentage(100);
-    console.log('ok'.green + ' -- Theme Processing Finished');
+    console.log(`${'ok'.green} -- Theme Processing Finished`);
 };
 
-utils.pollForJobCompletion = resultFilter => {
-    return async.retryable({
-        interval: 1000,
-        errorFilter: err => {
-            if (err.name === "JobCompletionStatusCheckPendingError") {
-                utils.markJobProgressPercentage(err.message);
-                return true;
-            }
+utils.pollForJobCompletion = (resultFilter) => {
+    return async.retryable(
+        {
+            interval: 1000,
+            errorFilter: (err) => {
+                if (err.name === 'JobCompletionStatusCheckPendingError') {
+                    utils.markJobProgressPercentage(err.message);
+                    return true;
+                }
 
-            return false;
+                return false;
+            },
+            times: Number.POSITIVE_INFINITY,
         },
-        times: Number.POSITIVE_INFINITY,
-    }, utils.checkIfJobIsComplete(resultFilter));
+        utils.checkIfJobIsComplete(resultFilter),
+    );
 };
 
-utils.checkIfJobIsComplete = resultFilter => (options, callback) => {
-    const config = options.config;
+utils.checkIfJobIsComplete = (resultFilter) => async (options) => {
+    const {
+        config: { accessToken },
+        apiHost,
+        storeHash,
+        bundleZipPath,
+        jobId,
+    } = options;
 
-    themeApiClient.getJob({
-        accessToken: config.accessToken,
-        apiHost: options.apiHost,
-        clientId: 'stencil-cli',
-        storeHash: options.storeHash,
-        bundleZipPath: options.bundleZipPath,
-        jobId: options.jobId,
+    const result = await themeApiClient.getJob({
+        accessToken,
+        apiHost,
+        storeHash,
+        bundleZipPath,
+        jobId,
         resultFilter,
-    }, (error, result) => {
-        if (error) {
-            return callback(error);
-        }
-
-        utils.markJobComplete();
-
-        callback(null, Object.assign({}, options, result));
     });
+
+    utils.markJobComplete();
+
+    return { ...options, ...result };
 };
 
-utils.promptUserWhetherToApplyTheme = (options, callback) => {
+utils.promptUserWhetherToApplyTheme = async (options) => {
     if (options.activate) {
-        callback(null, Object.assign({}, options, { applyTheme: true }));
-    } else {
-        const questions = [{
+        return { ...options, applyTheme: true };
+    }
+
+    const questions = [
+        {
             type: 'confirm',
             name: 'applyTheme',
-            message: `Would you like to apply your theme to your store at ${options.config.normalStoreUrl}?`,
+            message: `Would you like to apply your theme to your store?`,
             default: false,
-        }];
+        },
+    ];
+    const answers = await Inquirer.prompt(questions);
 
-        Inquirer.prompt(questions, answers => {
-            callback(null, Object.assign({}, options, { applyTheme: answers.applyTheme }));
-        });
-    }
+    return { ...options, ...answers };
 };
 
-utils.getVariations = (options, callback) => {
-    if (!options.applyTheme) {
-        return async.nextTick(callback.bind(null, null, options));
+utils.getChannels = async (options) => {
+    const {
+        config: { accessToken },
+        channelId,
+        apiHost,
+        storeHash,
+        applyTheme,
+    } = options;
+
+    if (!applyTheme || channelId) {
+        return options;
     }
 
-    themeApiClient.getVariationsByThemeId({
-        accessToken: options.accessToken,
-        apiHost: options.apiHost,
-        clientId: 'stencil-cli',
-        themeId: options.themeId,
-        storeHash: options.storeHash,
-    }, (error, result) => {
-        if (error) {
-            return callback(error);
-        }
-        if (options.activate !== true && options.activate !== undefined) {
-            const findVariation = result.variations.find(item => item.name === options.activate);
-            if (!findVariation || !findVariation.uuid) {
-                throw new Error(`Invalid theme variation provided!. Available options${result.variations.map(item => ` ${item.name}`).join(',')}...`);
-            }
-            callback(null, Object.assign({}, options, { variationId: findVariation.uuid }));
-        } else if (options.activate === true) {
-            callback(null, Object.assign({}, options, { variationId: result.variations[0].uuid }));
-        } else {
-            callback(null, Object.assign({}, options, result));
-        }
+    const channels = await themeApiClient.getStoreChannels({
+        accessToken,
+        apiHost,
+        storeHash,
     });
+
+    return { ...options, channels };
 };
 
-utils.promptUserForVariation = (options, callback) => {
-    if (!options.applyTheme) {
-        return async.nextTick(callback.bind(null, null, options));
+utils.getVariations = async (options) => {
+    const {
+        config: { accessToken },
+        apiHost,
+        storeHash,
+        themeId,
+        applyTheme,
+        activate,
+    } = options;
+
+    if (!applyTheme) {
+        return options;
     }
 
-    if (options.variationId) {
-        callback(null, options);
-    } else {
-        const questions = [{
+    const variations = await themeApiClient.getVariationsByThemeId({
+        accessToken,
+        apiHost,
+        themeId,
+        storeHash,
+    });
+
+    // Activate the default variation
+    if (activate === true) {
+        return { ...options, variationId: variations[0].uuid };
+    }
+    // Activate the specified variation
+    if (activate !== undefined) {
+        const foundVariation = variations.find((item) => item.name === activate);
+
+        if (!foundVariation || !foundVariation.uuid) {
+            const availableOptionsStr = variations.map((item) => `${item.name}`).join(', ');
+            throw new Error(
+                `Invalid theme variation provided! Available options: ${availableOptionsStr}.`,
+            );
+        }
+
+        return { ...options, variationId: foundVariation.uuid };
+    }
+
+    // Didn't specify a variation explicitly, will ask the user later
+    return { ...options, variations };
+};
+
+utils.promptUserForChannel = async (options) => {
+    const { applyTheme, channelId, channels } = options;
+
+    if (!applyTheme || channelId) {
+        return options;
+    }
+
+    const selectedChannelId = await utils.promptUserToSelectChannel(channels);
+    return { ...options, channelId: selectedChannelId };
+};
+
+utils.promptUserToSelectChannel = async (channels) => {
+    if (channels.length < 2) {
+        return channels[0].channel_id;
+    }
+
+    const questions = [
+        {
+            type: 'list',
+            name: 'channelId',
+            message: 'Which channel would you like to use?',
+            choices: channels.map((channel) => ({
+                name: channel.url,
+                value: channel.channel_id,
+            })),
+        },
+    ];
+
+    const answer = await Inquirer.prompt(questions);
+    return answer.channelId;
+};
+
+utils.promptUserForVariation = async (options) => {
+    if (!options.applyTheme || options.variationId) {
+        return options;
+    }
+
+    const questions = [
+        {
             type: 'list',
             name: 'variationId',
             message: 'Which variation would you like to apply?',
-            choices: options.variations.map(variation => ({ name: variation.name, value: variation.uuid })),
-        }];
+            choices: options.variations.map((variation) => ({
+                name: variation.name,
+                value: variation.uuid,
+            })),
+        },
+    ];
+    const answers = await Inquirer.prompt(questions);
 
-        Inquirer.prompt(questions, answers => {
-            callback(null, Object.assign({}, options, answers));
-        });
-    }
+    return { ...options, ...answers };
 };
 
 utils.requestToApplyVariationWithRetrys = () => {
-    return async.retryable({
-        interval: 1000,
-        errorFilter: err => {
-            if (err.name === "VariationActivationTimeoutError") {
-                console.log('warning'.yellow + ` -- Theme Activation Timed Out; Retrying...`);
-                return true;
-            }
+    return async.retryable(
+        {
+            interval: 1000,
+            errorFilter: (err) => {
+                if (err.name === 'VariationActivationTimeoutError') {
+                    console.log(`${'warning'.yellow} -- Theme Activation Timed Out; Retrying...`);
+                    return true;
+                }
 
-            return false;
+                return false;
+            },
+            times: 3,
         },
-        times: 3,
-    }, utils.requestToApplyVariation);
+        utils.requestToApplyVariation,
+    );
 };
 
-utils.requestToApplyVariation = (options, callback) => {
-    if (!options.applyTheme) {
-        return async.nextTick(callback.bind(null, null, options));
+utils.requestToApplyVariation = async (options) => {
+    const {
+        config: { accessToken },
+        apiHost,
+        storeHash,
+        variationId,
+        channelId,
+    } = options;
+
+    if (options.applyTheme) {
+        await themeApiClient.activateThemeByVariationId({
+            accessToken,
+            apiHost,
+            storeHash,
+            variationId,
+            channelId,
+        });
     }
 
-    themeApiClient.activateThemeByVariationId({
-        accessToken: options.accessToken,
-        apiHost: options.apiHost,
-        clientId: 'stencil-cli',
-        storeHash: options.storeHash,
-        variationId: options.variationId,
-    }, (error, result) => {
-        if (error) {
-            return callback(error);
-        }
-
-        callback(null, Object.assign({}, options, result));
-    });
+    return options;
 };
 
 utils.notifyUserOfCompletion = (options, callback) => {
     callback(null, 'Stencil Push Finished');
 };
+
+module.exports = utils;
