@@ -1,55 +1,42 @@
-'use strict';
+const axios = require('axios');
+const MockAdapter = require('axios-mock-adapter');
 
-const Code = require('code');
-const Lab = require('lab');
-const lab = exports.lab = Lab.script();
-const describe = lab.describe;
-const expect = Code.expect;
-const mockConfig = require('../test/_mocks/bin/dotStencilFile.json');
-const it = lab.it;
-const sinon = require('sinon');
-const Wreck = require('wreck');
-const utils = require('./stencil-push.utils');
-const mockDb = {
-    _data: {},
-    get data() {
-        return this._data;
-    },
-    set data(data) {
-        this._data = data;
-    },
-};
+const { getStoreHash } = require('./stencil-push.utils');
+
+const axiosMock = new MockAdapter(axios);
 
 describe('stencil push utils', () => {
-    lab.beforeEach(done => {
-        this.wreckReqStub = sinon.stub(Wreck, 'request').callsFake((method, url, options, callback) => {
-            callback(null, mockDb.data);
-        });
-        this.wreckReadStub = sinon.stub(Wreck, 'read').callsFake((response, options, callback) => {
-            callback(null, response);
-        });
-        done();
-    });
+    const mockConfig = {
+        normalStoreUrl: 'https://www.example.com',
+        port: 4000,
+        accessToken: 'accessTokenValue',
+    };
 
-    lab.afterEach(done => {
-        this.wreckReqStub.restore();
-        this.wreckReadStub.restore();
-        done();
+    afterEach(() => {
+        jest.restoreAllMocks();
+        axiosMock.reset();
     });
 
     describe('.getStoreHash()', () => {
-        mockDb.data = {
-            store_hash: 'abc123',
-            statusCode: 200,
-        };
+        it('should get the store hash', async () => {
+            const mockResponseData = {
+                store_hash: 'abc123',
+            };
+            axiosMock.onGet().reply(200, mockResponseData);
 
-        it('should get the store hash', done => {
-            utils.getStoreHash({ config: mockConfig }, (err, result) => {
-                expect(err).to.be.null();
-                expect(result.storeHash).to.be.equal(mockDb.data.store_hash);
-                expect(result.config.normalStoreUrl).to.be.equal(mockConfig.normalStoreUrl);
-                done();
-            });
+            const result = await getStoreHash({ config: mockConfig });
+
+            expect(result.storeHash).toEqual(mockResponseData.store_hash);
+            expect(result.config.normalStoreUrl).toEqual(mockConfig.normalStoreUrl);
+        });
+
+        it('should return an error if it fails to retrieve the store hash', async () => {
+            const mockResponseData = {};
+            axiosMock.onGet().reply(200, mockResponseData);
+
+            await expect(getStoreHash({ config: mockConfig })).rejects.toThrow(
+                'Received empty store_hash value in the server response',
+            );
         });
     });
 });
