@@ -1,14 +1,374 @@
-var _ = require('lodash');
-var Fs = require('fs');
-var Hoek = require('hoek');
-var Path = require('path');
-var Glob = require('glob');
-var Async = require('async');
-var jsonLint = require('./json-lint');
+const _ = require('lodash');
+const fs = require('fs');
+const Path = require('path');
+const glob = require('glob');
+const { promisify } = require('util');
+const { parseJsonFile } = require('./utils/fsUtils');
 
-var themeConfigInstance;
+class ThemeConfig {
+    /**
+     * ThemeConfig Class Constructor
+     *
+     * @param themePath
+     * @constructor
+     */
+    constructor(themePath) {
+        if (themePath) {
+            this.setThemePath(themePath);
+        }
 
-module.exports.getInstance = getInstance;
+        this.currentVariationSettings = null;
+        this.globalSettings = null;
+        this.variationIndex = 0;
+    }
+
+    /**
+     * Grab the config JSON string, parse it into an object, grab the current variation,
+     * and then merge it into the top level settings.
+     *
+     * @returns {object}
+     */
+    async getConfig() {
+        const config = await this.getRawConfig();
+        const variation = this._getVariation(config, this.variationIndex);
+
+        this.globalSettings = _.cloneDeep(config.settings);
+
+        if (!this.currentVariationSettings) {
+            this.currentVariationSettings = variation.settings;
+        }
+
+        // Set some defaults
+        config.css_compiler = config.css_compiler || 'scss';
+        config.autoprefixer_cascade = config.autoprefixer_cascade || true;
+        config.autoprefixer_browsers = config.autoprefixer_browsers || [
+            '> 1%',
+            'last 2 versions',
+            'Firefox ESR',
+        ];
+
+        // Merge in the variation settings and images objects
+        config.settings = _.defaultsDeep(this.currentVariationSettings, config.settings || {});
+        config.images = _.defaultsDeep(variation.images, config.images || {});
+        config.variationName = variation.name;
+
+        return config;
+    }
+
+    /**
+     * Updates the current variation settings
+     *
+     * @param newSettings
+     * @param saveToFile
+     * @returns {ThemeConfig}
+     */
+    async updateConfig(newSettings, saveToFile) {
+        let rawConfig;
+        let variation;
+
+        // Remove all variation settings that match theme's settings
+        this.currentVariationSettings = _.omit(
+            newSettings,
+            (value, key) => this.globalSettings[key] === value,
+            this,
+        );
+
+        if (saveToFile) {
+            // Resetting the global settings so the only data saved to the file are the changes
+            // from a specific variation's settings.
+            rawConfig = await this.getRawConfig();
+            variation = this._getVariation(rawConfig, this.variationIndex);
+
+            variation.settings = this.currentVariationSettings;
+
+            fs.writeFileSync(this.configPath, JSON.stringify(rawConfig, null, 2), {
+                encoding: 'utf-8',
+            });
+        }
+
+        return this;
+    }
+
+    /**
+     * Theme Path Setter
+     *
+     * @param themePath
+     * @returns {ThemeConfig}
+     */
+    setThemePath(themePath) {
+        this.themePath = themePath;
+        this.configPath = Path.join(themePath, 'config.json');
+        this.schemaPath = Path.join(themePath, 'schema.json');
+        this.schemaTranslationsPath = Path.join(themePath, 'schemaTranslations.json');
+        this.variationIndex = 0;
+
+        return this;
+    }
+
+    /**
+     * Reset Current Variation Settings
+     *
+     * @returns {ThemeConfig}
+     */
+    resetVariationSettings() {
+        this.currentVariationSettings = null;
+    }
+
+    /**
+     * Variation Name Setter
+     *
+     * @param variationIndex
+     * @returns {ThemeConfig}
+     */
+    setVariation(variationIndex) {
+        if (variationIndex !== this.variationIndex) {
+            this.variationIndex = variationIndex;
+            this.currentVariationSettings = null;
+        }
+
+        return this;
+    }
+
+    /**
+     * Get variation object by index
+     * @returns {Object}
+     */
+    async getVariation(variationIndex) {
+        const config = await this.getConfig();
+        return this._getVariation(config, variationIndex);
+    }
+
+    /**
+     * Variation Name Setter
+     *
+     * @param variationName
+     * @returns {ThemeConfig}
+     */
+    async setVariationByName(variationName) {
+        const config = await this.getRawConfig();
+        let variationIndex = 0;
+
+        if (variationName) {
+            variationIndex = config.variations.findIndex(
+                (variation) => variation.name === variationName,
+            );
+
+            if (variationIndex === -1) {
+                throw new Error(
+                    `Variation: ${variationName} is not defined in the theme's config.json`,
+                );
+            }
+        }
+
+        this.variationIndex = variationIndex;
+        this.currentVariationSettings = null;
+
+        return this;
+    }
+
+    /**
+     * Get theme name
+     * @returns {string}
+     */
+    async getName() {
+        const config = await this.getConfig();
+        return config.name || '';
+    }
+
+    /**
+     * Get theme version
+     * @returns {string}
+     */
+    async getVersion() {
+        const config = await this.getConfig();
+        return config.version || '';
+    }
+
+    /**
+     * Get the number of variations
+     * @returns {number}
+     */
+    async getVariationCount() {
+        const { variations } = await this.getConfig();
+
+        if (!Array.isArray(variations)) {
+            return 0;
+        }
+
+        return variations.length;
+    }
+
+    /**
+     * Check if the variation exists
+     * @returns {boolean}
+     */
+    async variationExists(variationIndex) {
+        return _.inRange(variationIndex, 0, await this.getVariationCount());
+    }
+
+    /**
+     * Check if the config.json file exists
+     * @returns {boolean}
+     */
+    configExists() {
+        return fs.existsSync(this.configPath);
+    }
+
+    /**
+     * Check if the schema.json file exists
+     * @returns {boolean}
+     */
+    schemaExists() {
+        return fs.existsSync(this.schemaPath);
+    }
+
+    /**
+     * Check if the schemaTranslations.json file exists
+     * @returns {boolean}
+     */
+    schemaTranslationsExists() {
+        return fs.existsSync(this.schemaTranslationsPath);
+    }
+
+    /**
+     * Scans the theme template directory for theme settings that need force reload
+     *
+     * @returns {Object}
+     */
+    async getSchema() {
+        if (!fs.existsSync(this.schemaPath)) {
+            return [];
+        }
+
+        const themeSchema = await parseJsonFile(this.schemaPath);
+
+        if (!Array.isArray(themeSchema)) {
+            return [];
+        }
+
+        const files = await promisify(glob)(Path.join(this.themePath, 'templates/**/*.html'));
+
+        const themeSettings = await Promise.all(files.map(this._fetchThemeSettings));
+
+        const forceReloadIds = themeSettings.reduce((res, item) => ({ ...res, ...item }), {});
+
+        for (const themeSchemaItem of themeSchema) {
+            for (const item of themeSchemaItem.settings) {
+                if (forceReloadIds[item.id]) {
+                    item.force_reload = true;
+                }
+            }
+        }
+
+        return themeSchema;
+    }
+
+    /**
+     * Get Schema Translations
+     *
+     * @returns {Promise<Object>}
+     */
+    async getSchemaTranslations() {
+        if (!fs.existsSync(this.schemaTranslationsPath)) {
+            return {};
+        }
+
+        const schemaTranslations = await parseJsonFile(this.schemaTranslationsPath);
+
+        return _.isObject(schemaTranslations) ? schemaTranslations : {};
+    }
+
+    /**
+     * Returns the parsed config.json data.
+     * In contrast with getConfig which returns merged settings.
+     *
+     * @returns {Promise<Object>}
+     */
+    async getRawConfig() {
+        return parseJsonFile(this.configPath);
+    }
+
+    /**
+     * Returns the parsed schema.json data
+     *
+     * @returns {Promise<Object>}
+     */
+    async getRawSchema() {
+        return parseJsonFile(this.schemaPath);
+    }
+
+    /**
+     * Returns the parsed schemaTranslations.json data
+     *
+     * @returns {Promise<Object>}
+     */
+    async getRawSchemaTranslations() {
+        return parseJsonFile(this.schemaTranslationsPath);
+    }
+
+    /**
+     * Grabs out a variation based on a name. Or if the name is not passed in, the very first one in the list.
+     *
+     * @private
+     * @param config
+     * @param variationIndex
+     * @returns {object}
+     */
+    _getVariation(config, variationIndex) {
+        let variation;
+
+        if (!Array.isArray(config.variations) || config.variations.length === 0) {
+            throw new Error('Your theme must have at least one variation in the config.json file.');
+        }
+
+        if (!variationIndex) {
+            [variation] = config.variations;
+        } else {
+            variation = config.variations[variationIndex];
+
+            if (!variation) {
+                throw new Error(`Variation index: ${variationIndex} not found theme's config.json`);
+            }
+        }
+
+        if (!_.isObject(variation.settings)) {
+            variation.settings = {};
+        }
+
+        if (!_.isObject(variation.images)) {
+            variation.images = {};
+        }
+
+        if (!_.isObject(variation.meta)) {
+            variation.meta = {};
+        }
+
+        return variation;
+    }
+
+    /**
+     * Scan file for theme_settings.*
+     * @private
+     * @param  {string}   path
+     * @returns {Object}
+     */
+    async _fetchThemeSettings(path) {
+        const themeSettingsRegexPattern = /\Wtheme_settings\.(.+?)\W/g;
+        const themeSettings = {};
+        let match;
+
+        const content = await fs.promises.readFile(path, 'utf8');
+
+        // eslint-disable-next-line
+        while ((match = themeSettingsRegexPattern.exec(content))) {
+            themeSettings[match[1]] = true;
+        }
+
+        return themeSettings;
+    }
+}
+
+let themeConfigInstance;
 
 /**
  * Grabs the currently set instance or creates one if it doesn't exist.  If called
@@ -31,464 +391,6 @@ function getInstance(themePath) {
     return themeConfigInstance;
 }
 
-/**
- * ThemeConfig Class Constructor
- *
- * @param themePath
- * @constructor
- */
-function ThemeConfig(themePath) {
-    if (themePath) {
-        this.setThemePath(themePath);
-    }
-
-    this.currentVariationSettings = null;
-    this.globalSettings = null;
-    this.variationIndex = 0;
-}
-
-/**
- * Grab the config JSON string, parse it into an object, grab the current variation,
- * and then merge it into the top level settings.
- *
- * @returns {object}
- */
-ThemeConfig.prototype.getConfig = function () {
-    var config = this.getRawConfig();
-    var variation = getVariation(config, this.variationIndex);
-
-    this.globalSettings = Hoek.clone(config.settings);
-
-    if (!this.currentVariationSettings) {
-        this.currentVariationSettings = variation.settings;
-    }
-
-    // Set some defaults
-    config.css_compiler = config.css_compiler || 'scss';
-    config.autoprefixer_cascade = config.autoprefixer_cascade || true;
-    config.autoprefixer_browsers = config.autoprefixer_browsers || ['> 1%', 'last 2 versions', 'Firefox ESR'];
-
-    // Merge in the variation settings and images objects
-    config.settings = Hoek.applyToDefaults(config.settings || {}, this.currentVariationSettings);
-    config.images = Hoek.applyToDefaults(config.images || {}, variation.images);
-    config.variationName = variation.name;
-
-    return config;
+module.exports = {
+    getInstance,
 };
-
-/**
- * Updates the current variation settings
- *
- * @param newSettings
- * @param saveToFile
- * @returns {ThemeConfig}
- */
-ThemeConfig.prototype.updateConfig = function (newSettings, saveToFile) {
-    var rawConfig;
-    var variation;
-
-    // Remove all variation settings that match theme's settings
-    this.currentVariationSettings = _.omit(newSettings, function (value, key) {
-        return this.globalSettings[key] === value;
-    }, this);
-
-    if (saveToFile) {
-        // Resetting the global settings so the only data saved to the file are the changes
-        // from a specific variation's settings.
-        rawConfig = this.getRawConfig();
-        variation = getVariation(rawConfig, this.variationIndex);
-
-        variation.settings = this.currentVariationSettings;
-
-        Fs.writeFileSync(this.configPath, JSON.stringify(rawConfig, null, 2), {encoding: 'utf-8'});
-    }
-
-    return this;
-};
-
-/**
- * Theme Path Setter
- *
- * @param themePath
- * @returns {ThemeConfig}
- */
-ThemeConfig.prototype.setThemePath = function (themePath) {
-    this.themePath = themePath;
-    this.configPath = Path.join(themePath, 'config.json');
-    this.schemaPath = Path.join(themePath, 'schema.json');
-    this.schemaTranslationsPath = Path.join(themePath, 'schemaTranslations.json');
-    this.variationIndex = 0;
-
-    return this;
-};
-
-/**
- * Variation Name Setter
- *
- * @param variationIndex
- * @returns {ThemeConfig}
- */
-ThemeConfig.prototype.setVariation = function (variationIndex) {
-
-    if (variationIndex !== this.variationIndex) {
-        this.variationIndex = variationIndex;
-        this.currentVariationSettings = null;
-    }
-
-    return this;
-};
-
-/**
- * Get the current variation object
- * @return {Object}
- */
-ThemeConfig.prototype.getCurrentVariation = function () {
-    if (_.isNull(this.variationIndex)) {
-        throw new Error('The current varation has not been set yet');
-    }
-
-    return getVariation(this.getConfig(), this.variationIndex);
-};
-
-
-/**
- * Get variation object by index
- * @return {Object}
- */
-ThemeConfig.prototype.getVariation = function (variationIndex) {
-    return getVariation(this.getConfig(), variationIndex);
-};
-
-/**
- * Variation Name Setter
- *
- * @param variationName
- * @returns {ThemeConfig}
- */
-ThemeConfig.prototype.setVariationByName = function (variationName) {
-    var config = this.getRawConfig();
-    var variationIndex = 0;
-
-    if (variationName) {
-        variationIndex = _.findIndex(config.variations, function (variation) {
-            return variation.name == variationName;
-        });
-
-        if (variationIndex === -1) {
-            throw new Error(
-                'Variation: ' +
-                variationName +
-                ' is not defined in the theme\'s config.json',
-            );
-        }
-    }
-
-    this.variationIndex = variationIndex;
-    this.currentVariationSettings = null;
-
-    return this;
-};
-
-/**
- * Get theme meta object
- * @return {Object}
- */
-ThemeConfig.prototype.getMeta = function () {
-    return this.getConfig().meta || {};
-};
-
-/**
- * Get theme name
- * @return {String}
- */
-ThemeConfig.prototype.getName = function () {
-    return this.getConfig().name || '';
-};
-
-/**
- * Get theme description
- * @return {String}
- */
-ThemeConfig.prototype.getDescription = function () {
-    return this.getConfig().description || '';
-};
-
-/**
- * Get theme version
- * @return {String}
- */
-ThemeConfig.prototype.getVersion = function () {
-    return this.getConfig().version || '';
-};
-
-/**
- * Get theme variation name
- * @return {String}
- */
-ThemeConfig.prototype.getVariationName = function () {
-    return this.getConfig().variationName || '';
-};
-
-/**
- * Get the number of variations
- * @return {Number}
- */
-ThemeConfig.prototype.getVariationCount = function () {
-    var variations = this.getConfig().variations;
-
-    if (!_.isArray(variations)) {
-        return 0;
-    }
-
-    return variations.length;
-};
-
-/**
- * Check if the variation exists
- * @return {Boolena}
- */
-ThemeConfig.prototype.variationExists = function (variationIndex) {
-    return _.inRange(variationIndex, 0, this.getVariationCount());
-};
-
-/**
- * Get theme price
- * @return {Number}
- */
-ThemeConfig.prototype.getPrice = function () {
-    return this.getMeta().price || 0;
-};
-
-/**
- * Get theme composed image
- * @return {String}
- */
-ThemeConfig.prototype.getComposedImage = function () {
-    return this.getMeta().composed_image || '';
-};
-
-/**
- * Get theme demo url
- * @return {String}
- */
-ThemeConfig.prototype.getDemoUrl = function () {
-    return this.getConfig().demoUrl || '';
-};
-
-/**
- * Check if the config.json file exists
- * @return {Boolean}
- */
-ThemeConfig.prototype.configExists = function () {
-    return fileExists(this.configPath);
-};
-
-/**
- * Check if the schema.json file exists
- * @return {Boolean}
- */
-ThemeConfig.prototype.schemaExists = function () {
-    return fileExists(this.schemaPath);
-};
-
-/**
- * Check if the schemaTranslations.json file exists
- * @return {Boolean}
- */
-ThemeConfig.prototype.schemaTranslationsExists = function () {
-    return fileExists(this.schemaTranslationsPath);
-};
-
-/**
- * Scans the theme template directory for theme settings that need force reload
- *
- * @param {Function} callback
- */
-ThemeConfig.prototype.getSchema = function (callback) {
-    var themeSchemaContent;
-    var themeSchema;
-
-    try {
-        themeSchemaContent = Fs.readFileSync(this.schemaPath, {encoding: 'utf-8'});
-    } catch (err) {
-        return callback(null, []);
-    }
-
-    if (themeSchemaContent) {
-        try {
-            themeSchema = jsonLint.parse(themeSchemaContent, this.schemaPath);
-        } catch (err) {
-            return callback(err);
-        }
-    }
-
-    if (!_.isArray(themeSchema)) {
-        themeSchema = [];
-    }
-
-    Glob(Path.join(this.themePath, 'templates/**/*.html'), function (err, files) {
-        if (err) {
-            return callback(err);
-        }
-
-        Async.map(files, fetchThemeSettings, function (err, themeSettings) {
-            var forceReloadIds = {};
-
-            if (err) {
-                return callback(err);
-            }
-
-            _.each(themeSettings, function (id) {
-                forceReloadIds = _.merge(forceReloadIds, id);
-            });
-
-            _.each(themeSchema, function (data) {
-                _.each(data.settings, function (item) {
-                    if (forceReloadIds[item.id]) {
-                        item['force_reload'] = true;
-                    }
-                });
-            });
-
-            return callback(null, themeSchema);
-        });
-    });
-};
-
-/**
- * Scan file for theme_settings.*
- * @param  {String}   path
- * @param  {Function} callback
- */
-function fetchThemeSettings(path, callback) {
-    var themeSettingsRegexPattern = /\Wtheme_settings\.(.+?)\W/g;
-    var themeSettings = {};
-
-    Fs.readFile(path, 'utf8', function (err, content) {
-        var match;
-
-        if (err) {
-            return callback(err);
-        }
-
-        while (match = themeSettingsRegexPattern.exec(content)) {
-            themeSettings[match[1]] = true;
-        }
-
-        return callback(null, themeSettings);
-    });
-};
-
-/**
- * Get Schema Translations
- *
- * @param {Function} callback
- */
-ThemeConfig.prototype.getSchemaTranslations = function(callback) {
-    var schemaTranslationsContent;
-    var schemaTranslations;
-
-    try {
-        schemaTranslationsContent = Fs.readFileSync(this.schemaTranslationsPath, {encoding: 'utf-8'});
-    } catch (err) {
-        return callback(null, {});
-    }
-
-    if (schemaTranslationsContent) {
-        try {
-            schemaTranslations = jsonLint.parse(schemaTranslationsContent, this.schemaTranslationsPath);
-        } catch (err) {
-            return callback(err);
-        }
-    }
-
-    if (!_.isObject(schemaTranslations)) {
-        schemaTranslations = {};
-    }
-
-    return callback(null, schemaTranslations);
-};
-
-/**
- * Return the raw config.json data
- *
- * @return {object}
- */
-ThemeConfig.prototype.getRawConfig = function() {
-    return jsonLint.parse(Fs.readFileSync(this.configPath, {encoding: 'utf-8'}), this.configPath);
-};
-
-/**
- * Return the raw schema.json data
- *
- * @return {object}
- */
-ThemeConfig.prototype.getRawSchema = function() {
-    return jsonLint.parse(Fs.readFileSync(this.schemaPath, {encoding: 'utf-8'}), this.schemaPath);
-};
-
-/**
- * Return the raw config.json data
- *
- * @return {object}
- */
-ThemeConfig.prototype.getRawSchemaTranslations = function() {
-    return jsonLint.parse(Fs.readFileSync(this.schemaTranslationsPath, {encoding: 'utf-8'}), this.schemaTranslationsPath);
-};
-
-/**
- * Grabs out a variation based on a name. Or if the name is not passed in, the very first one in the list.
- *
- * @param config
- * @param variationIndex
- * @returns {object}
- */
-function getVariation(config, variationIndex) {
-    var variation;
-
-    if (! _.isArray(config.variations) || config.variations.length === 0) {
-        throw new Error(
-            'Your theme must have at least one variation in the config.json file.',
-        );
-    }
-
-    if (! variationIndex) {
-        variation = config.variations[0];
-    } else {
-        variation = config.variations[variationIndex];
-
-        if (! variation) {
-            throw new Error(
-                'Variation index: ' + variationIndex + ' not found theme\'s config.json',
-            );
-        }
-    }
-
-    if (!_.isObject(variation.settings)) {
-        variation.settings = {};
-    }
-
-    if (!_.isObject(variation.images)) {
-        variation.images = {};
-    }
-
-    if (!_.isObject(variation.meta)) {
-        variation.meta = {};
-    }
-
-    return variation;
-}
-
-/**
- * Check if file exist syncronous
- * @param  {string} path
- * @return {boolean}
- */
-function fileExists(path) {
-    try {
-        return !!Fs.statSync(path);
-    }
-    catch (e) {
-        return false;
-    }
-}
