@@ -1,143 +1,136 @@
-'use strict';
+const _ = require('lodash');
+const path = require('path');
+const os = require('os');
+const fs = require('fs');
 
-const Code = require('code');
-const Os = require('os');
-const Fs = require('fs');
-const Hoek = require('hoek');
-const Lab = require('lab');
-const Path = require('path');
-const Sinon = require('sinon');
 const ThemeConfig = require('./theme-config');
-const lab = exports.lab = Lab.script();
-const themePath = Path.join(process.cwd(), 'test/_mocks/themes/valid');
-const missingVariationsThemePath = Path.join(process.cwd(), 'test/_mocks/themes/missing-variation');
-const bareBonesThemePath = Path.join(process.cwd(), 'test/_mocks/themes/bare-bones');
-const describe = lab.describe;
-const expect = Code.expect;
-const it = lab.it;
+const { parseJsonFile } = require('./utils/fsUtils');
+const originalSchema = require('../test/_mocks/themes/valid/schema.json');
+
+const themePath = path.join(process.cwd(), 'test/_mocks/themes/valid');
+const missingVariationsThemePath = path.join(process.cwd(), 'test/_mocks/themes/missing-variation');
+const bareBonesThemePath = path.join(process.cwd(), 'test/_mocks/themes/bare-bones');
 
 describe('ThemeConfig', () => {
-    let themeConfig;
-
-    lab.beforeEach(done => {
-        themeConfig = ThemeConfig.getInstance(themePath).setVariationByName('First');
-
-        done();
+    afterEach(async () => {
+        jest.restoreAllMocks();
     });
 
     describe('getInstance()', () => {
-        it ('should always return the same instance', done => {
-            expect(themeConfig).to.equal(ThemeConfig.getInstance());
+        it('should always return the same instance', async () => {
+            const themeConfig = await ThemeConfig.getInstance(themePath);
 
-            done();
+            expect(themeConfig).toBe(ThemeConfig.getInstance());
         });
 
-        it ('should allow overwriting of configPath, schemaPath, and variationName by calling getInstance with params', done => {
-            const isWindows = Os.platform() === 'win32';
+        it('should allow overwriting of configPath, schemaPath, and variationName by calling getInstance with params', async () => {
+            const themeConfig = ThemeConfig.getInstance(themePath);
+            await themeConfig.setVariationByName('First');
+
+            const isWindows = os.platform() === 'win32';
             const secondConfigPath = isWindows ? '\\fake\\config.json' : '/fake/config.json';
             const secondSchemaPath = isWindows ? '\\fake\\schema.json' : '/fake/schema.json';
             const secondThemeConfig = ThemeConfig.getInstance('/fake');
 
-            expect(secondThemeConfig.configPath).to.equal(secondConfigPath);
-            expect(secondThemeConfig.schemaPath).to.equal(secondSchemaPath);
-            expect(secondThemeConfig).to.equal(themeConfig);
-
-            done();
+            expect(secondThemeConfig.configPath).toEqual(secondConfigPath);
+            expect(secondThemeConfig.schemaPath).toEqual(secondSchemaPath);
+            expect(secondThemeConfig).toEqual(themeConfig);
         });
     });
 
     describe('getConfig()', () => {
-        it('should return the correct config for the current variation', done => {
-            var config = ThemeConfig.getInstance().setVariationByName('Second').getConfig(),
-                originalSettingsToCompare = {
-                    color: '#ffffff',
-                    font: 'Sans Something',
-                    select: 'first',
-                    checkbox: true,
-                    radio: 'first',
-                },
-                settingsToCompare = {
-                    color: '#000000',
-                    font: 'Sans Something',
-                    select: 'second',
-                    checkbox: true,
-                    radio: 'first',
-                },
-                imagesToCompare = {
-                    logo: { width: 200, height: 200 },
-                    thumb: { width: 10, height: 10 },
-                };
+        it('should return the correct config for the current variation', async () => {
+            const themeConfig = ThemeConfig.getInstance(themePath);
+            await themeConfig.setVariationByName('Second');
 
-            expect(config.settings).to.equal(settingsToCompare);
-            expect(config.images).to.equal(imagesToCompare);
-            expect(ThemeConfig.getInstance().globalSettings).to.equal(originalSettingsToCompare);
+            const config = await themeConfig.getConfig();
+            const originalSettingsToCompare = {
+                color: '#ffffff',
+                font: 'Sans Something',
+                select: 'first',
+                checkbox: true,
+                radio: 'first',
+            };
+            const settingsToCompare = {
+                color: '#000000',
+                font: 'Sans Something',
+                select: 'second',
+                checkbox: true,
+                radio: 'first',
+            };
+            const imagesToCompare = {
+                logo: { width: 200, height: 200 },
+                thumb: { width: 10, height: 10 },
+            };
 
-            done();
+            expect(config.settings).toEqual(settingsToCompare);
+            expect(config.images).toEqual(imagesToCompare);
+            expect(ThemeConfig.getInstance().globalSettings).toEqual(originalSettingsToCompare);
         });
 
-        it('should throw an Error if there are no variations in config.json file', done => {
-            var themeConfig = ThemeConfig.getInstance(missingVariationsThemePath);
+        it('should throw an Error if there are no variations in config.json file', async () => {
+            const themeConfig = ThemeConfig.getInstance(missingVariationsThemePath);
 
-            expect(themeConfig.getConfig).to.throw(Error);
-
-            done();
+            await expect(themeConfig.getConfig).rejects.toThrow(Error);
         });
 
-        it('should grab the first variation if none is passed in', done => {
-            var themeConfig = ThemeConfig.getInstance();
+        it('should grab the first variation if none is passed in', async () => {
+            const themeConfig = ThemeConfig.getInstance(themePath);
 
-            themeConfig.setVariationByName(null);
-            // console.log(themeConfig.getConfig());
-            expect(themeConfig.getConfig().variationName).to.equal('First');
+            await themeConfig.setVariationByName(null);
 
-            done();
+            const config = await themeConfig.getConfig();
+            expect(config.variationName).toEqual('First');
         });
 
-        it('should grab a specific variation if passed in', done => {
-            var themeConfig = ThemeConfig.getInstance();
+        it('should grab a specific variation if passed in', async () => {
+            const themeConfig = ThemeConfig.getInstance(themePath);
 
-            themeConfig.setVariationByName('Second');
-            expect(themeConfig.getConfig().variationName).to.equal('Second');
+            await themeConfig.setVariationByName('Second');
 
-            done();
+            const config = await themeConfig.getConfig();
+            expect(config.variationName).toEqual('Second');
         });
 
-        it('should throw an Error if the passed in variation name does not match any in the config.json', done => {
-            var themeConfig = ThemeConfig.getInstance(themePath);
+        it('should throw an Error if the passed in variation name does not match any in the config.json', async () => {
+            const themeConfig = ThemeConfig.getInstance(themePath);
 
-            function setVariation() {
-                themeConfig.setVariationByName('Does Not Exist');
+            async function setVariation() {
+                await themeConfig.setVariationByName('Does Not Exist');
             }
 
-            expect(setVariation).to.throw(Error);
-
-            done();
+            await expect(setVariation).rejects.toThrow(Error);
         });
 
-        it('should set proper default values if they do not exist', done => {
-            var themeConfig = ThemeConfig.getInstance(bareBonesThemePath),
-                config = themeConfig.getConfig();
+        it('should set proper default values if they do not exist', async () => {
+            const themeConfig = ThemeConfig.getInstance(bareBonesThemePath);
+            await themeConfig.setVariationByName('First');
 
-            expect(config.settings).to.equal({});
-            expect(config.images).to.equal({});
-            expect(config.css_compiler).to.equal('scss');
-            expect(config.autoprefixer_cascade).to.equal(true);
-            expect(config.autoprefixer_browsers).to.equal(['> 1%', 'last 2 versions', 'Firefox ESR']);
+            const config = await themeConfig.getConfig();
 
-            done();
+            expect(config.settings).toEqual({});
+            expect(config.images).toEqual({});
+            expect(config.css_compiler).toEqual('scss');
+            expect(config.autoprefixer_cascade).toEqual(true);
+            expect(config.autoprefixer_browsers).toEqual([
+                '> 1%',
+                'last 2 versions',
+                'Firefox ESR',
+            ]);
         });
     });
 
     describe('updateConfig()', () => {
-        var config,
-            newSettings,
-            originalSettings,
-            themeConfig;
+        let newSettings;
+        let originalSettings;
+        let themeConfig;
 
-        lab.beforeEach(done => {
-            themeConfig = ThemeConfig.getInstance();
-            config = themeConfig.getConfig();
-            originalSettings = Hoek.clone(config.settings);
+        beforeEach(async () => {
+            themeConfig = ThemeConfig.getInstance(themePath);
+            await themeConfig.setVariationByName('First');
+
+            const config = await themeConfig.getConfig();
+            originalSettings = _.cloneDeep(config.settings);
             newSettings = {
                 color: '#000000',
                 font: 'Sans Something',
@@ -145,100 +138,77 @@ describe('ThemeConfig', () => {
                 checkbox: true,
                 radio: 'first',
             };
-
-            done();
         });
 
-        it('should update the currentConfig with the new changes', done => {
-            expect(originalSettings).not.to.equal(newSettings);
+        it('should update the currentConfig with the new changes', async () => {
+            expect(originalSettings).not.toEqual(newSettings);
 
-            themeConfig.updateConfig(newSettings);
+            await themeConfig.updateConfig(newSettings);
 
-            expect(themeConfig.getConfig().settings).not.to.equal(originalSettings);
-            expect(themeConfig.getConfig().settings).to.equal(newSettings);
-
-            done();
+            const config = await themeConfig.getConfig();
+            expect(config.settings).not.toEqual(originalSettings);
+            expect(config.settings).toEqual(newSettings);
         });
 
-        it('should not write to config file if saveToFile is falsey', done => {
-            Sinon.stub(Fs, 'writeFileSync');
+        it('should not write to config file if saveToFile is falsy', async () => {
+            jest.spyOn(fs, 'writeFileSync').mockImplementation(jest.fn());
 
-            themeConfig.updateConfig(newSettings);
-            expect(Fs.writeFileSync.called).to.be.false();
-            Fs.writeFileSync.restore();
+            await themeConfig.updateConfig(newSettings);
 
-            done();
+            expect(fs.writeFileSync.called).toBeFalsy();
         });
 
-        it('should write to config file if saveToFile is truthy', done => {
-            Sinon.stub(Fs, 'writeFileSync');
+        it('should write to config file if saveToFile is truthy', async () => {
+            jest.spyOn(fs, 'writeFileSync').mockImplementation(jest.fn());
 
-            themeConfig.updateConfig(newSettings, true);
-            expect(Fs.writeFileSync.called).to.be.true();
-            Fs.writeFileSync.restore();
+            await themeConfig.updateConfig(newSettings, true);
 
-            done();
+            expect(fs.writeFileSync).toHaveBeenCalled();
         });
 
-        it('should just modify the variations section of the file', done => {
-            var configPath = Path.join(themePath, 'config.json');
-            var initialConfig = JSON.parse(Fs.readFileSync(configPath, {encoding: 'utf-8'}));
+        it('should just modify the variations section of the file', async () => {
+            const configPath = path.join(themePath, 'config.json');
+            const initialConfig = await parseJsonFile(configPath);
 
-            Sinon.stub(Fs, 'writeFileSync').callsFake(testSave);
+            const writeFileSyncStub = jest.spyOn(fs, 'writeFileSync').mockImplementation(jest.fn());
 
-            function testSave(path, newConfig) {
-                newConfig = JSON.parse(newConfig);
+            await themeConfig.updateConfig(newSettings, true);
 
-                expect(newConfig).not.to.equal(initialConfig);
+            const newConfig = JSON.parse(writeFileSyncStub.mock.calls[0][1]);
 
-                delete newConfig.variations;
-                delete initialConfig.variations;
+            expect(newConfig).not.toEqual(initialConfig);
 
-                expect(newConfig).to.equal(initialConfig);
+            delete newConfig.variations;
+            delete initialConfig.variations;
 
-                done();
-            }
-
-            themeConfig.updateConfig(newSettings, true);
-
-            Fs.writeFileSync.restore();
+            expect(newConfig).toEqual(initialConfig);
         });
     });
 
     describe('getSchema()', () => {
+        it('should return the correct schema', async () => {
+            const themeConfig = ThemeConfig.getInstance(themePath);
+            await themeConfig.setVariationByName('First');
 
-        it('should return the correct schema', done => {
+            const schema = await themeConfig.getSchema();
 
-            var themeConfig = ThemeConfig.getInstance();
-            var originalSchema = require(Path.join(themePath, 'schema.json'));
+            expect(schema).toBeInstanceOf(Array);
 
-            themeConfig.getSchema((err, schema) => {
+            expect(schema[0]).toEqual(originalSchema[0]);
+            expect(schema).toHaveLength(2);
 
-                expect(err).to.be.null();
-
-                expect(schema).to.be.an.array();
-
-                expect(schema[0]).to.equal(originalSchema[0]);
-                expect(schema).to.have.length(2);
-
-                expect(schema[1].settings[0].force_reload).to.true();
-                expect(schema[1].settings[1].force_reload).to.true();
-                expect(schema[1].settings[2].force_reload).to.true();
-
-                done();
-            });
+            expect(schema[1].settings[0].force_reload).toBeTruthy();
+            expect(schema[1].settings[1].force_reload).toBeTruthy();
+            expect(schema[1].settings[2].force_reload).toBeTruthy();
         });
 
-        it('should return an empty schema', done => {
+        it('should return an empty schema', async () => {
+            const themeConfig = ThemeConfig.getInstance(bareBonesThemePath);
 
-            var themeConfig = ThemeConfig.getInstance(bareBonesThemePath);
+            const schema = await themeConfig.getSchema();
 
-            themeConfig.getSchema((err, schema) => {
-                expect(schema).to.be.an.array();
-                expect(schema).to.have.length(0);
-
-                done();
-            });
+            expect(schema).toBeInstanceOf(Array);
+            expect(schema).toHaveLength(0);
         });
     });
 });
