@@ -1,7 +1,9 @@
+const schemaTranslations = require('../schemas/schemaTranslations.json');
+
 class ValidatorSchemaTranslations {
     constructor() {
         this.trackedKeys = ['name', 'content', 'label', 'settings', 'options', 'group'];
-        this.validationSchema = require('../schemas/schemaTranslations.json');
+        this.validationSchema = schemaTranslations;
         this.translations = {};
         this.schemaKeys = [];
         this.translationsKeys = [];
@@ -37,7 +39,7 @@ class ValidatorSchemaTranslations {
 
     /**
      * Get validation schema
-     * @return {object}
+     * @returns {object}
      */
     getValidationSchema() {
         return this.validationSchema;
@@ -45,7 +47,7 @@ class ValidatorSchemaTranslations {
 
     /**
      * Get translations
-     * @return {array}
+     * @returns {array}
      */
     getTranslations() {
         return this.translations;
@@ -53,7 +55,7 @@ class ValidatorSchemaTranslations {
 
     /**
      * Get i18n keys from translations
-     * @return {array}
+     * @returns {array}
      */
     getTranslationsKeys() {
         this.translationsKeys = Object.keys(this.translations);
@@ -61,7 +63,7 @@ class ValidatorSchemaTranslations {
 
     /**
      * Get i18n keys from schema
-     * @return {array}
+     * @returns {array}
      */
     getSchemaKeys() {
         return this.schemaKeys;
@@ -69,43 +71,46 @@ class ValidatorSchemaTranslations {
 
     /**
      * Get translatable strings
-     * @param {array} schema
+     * @param {object[]} schema
      */
     getTranslatableStrings(schema) {
-        const context = this;
-
-        schema.forEach(element => {
-            Object.entries(element).forEach(function (item) {
-                const key = item[0];
-                const value = item[1];
-
-                if (!context.trackedKeys.includes(key)) {
-                    return;
+        for (const element of schema) {
+            for (const [key, value] of Object.entries(element)) {
+                if (!this.trackedKeys.includes(key)) {
+                    continue;
                 }
 
                 if (Array.isArray(value)) {
-                    return context.getTranslatableStrings(value);
+                    this.getTranslatableStrings(value);
                 }
 
-                context.setSchemaKeys(value);
-            });
-        });
+                this.setSchemaKeys(value);
+            }
+        }
     }
 
     /**
      * Find unused i18n keys in schemaTranslations
-     * @return {array}
+     * @returns {array}
      */
     findUnusedKeys() {
-        return this.translationsKeys.filter(key => !this.schemaKeys.includes(key));
+        return this.translationsKeys.filter((key) => {
+            const translationRegex = /^i18n.RegionName.*?/g;
+            const translationMatch = translationRegex.exec(key);
+            if (translationMatch) {
+                return false;
+            }
+
+            return !this.schemaKeys.includes(key);
+        });
     }
 
     /**
      * Find missed i18n keys in schemaTranslations
-     * @return {array}
+     * @returns {array}
      */
     findMissedKeys() {
-        return this.schemaKeys.filter(key => !this.translationsKeys.includes(key));
+        return this.schemaKeys.filter((key) => !this.translationsKeys.includes(key));
     }
 }
 
