@@ -1,67 +1,49 @@
-const Paths = require('path');
-const Code = require('code');
-const Lab = require('lab');
-const lab = exports.lab = Lab.script();
-const describe = lab.describe;
-const it = lab.it;
-const expect = Code.expect;
-
 const ValidatorSchemaTranslations = require('./schema-translations');
-const schema = require(Paths.join(process.cwd(), 'test/_mocks/themes/valid/schema.json'));
-const schemaTranslations = require(Paths.join(process.cwd(), 'test/_mocks/themes/valid/schemaTranslations.json'));
-const validationsTranslations = require('../schemas/schemaTranslations');
+const schema = require('../../test/_mocks/themes/valid/schema.json');
+const schemaTranslations = require('../../test/_mocks/themes/valid/schemaTranslations.json');
+const validationsTranslations = require('../schemas/schemaTranslations.json');
 
 const validatorSchemaTranslations = () => {
     return new ValidatorSchemaTranslations();
 };
 
 describe('ValidatorSchemaTranslations', () => {
-    it('should return translations', function (done) {
+    it('should return translations', () => {
         const instance = validatorSchemaTranslations();
 
         instance.setTranslations(schemaTranslations);
 
-        expect(instance.getTranslations()).equals(schemaTranslations);
-
-        done();
+        expect(instance.getTranslations()).toEqual(schemaTranslations);
     });
 
-    it('should return translations keys', function (done) {
+    it('should return translations keys', () => {
         const instance = validatorSchemaTranslations();
 
         instance.setSchema(schema);
 
-        expect(instance.getSchemaKeys()).equals(['i18n.Test']);
-
-        done();
+        expect(instance.getSchemaKeys()).toEqual(['i18n.Test']);
     });
 
-    it('should return validation schema', function (done) {
+    it('should return validation schema', () => {
         const instance = validatorSchemaTranslations();
 
-        expect(instance.getValidationSchema()).equals(validationsTranslations);
-
-        done();
+        expect(instance.getValidationSchema()).toEqual(validationsTranslations);
     });
 
-    it('should return i18n keys array without duplicates', function (done) {
+    it('should return i18n keys array without duplicates', () => {
         const instance = validatorSchemaTranslations();
 
         instance.setSchemaKeys('i18n.Global');
         instance.setSchemaKeys('i18n.Global');
 
-        expect(instance.getSchemaKeys()).equals(['i18n.Global']);
-
-        done();
+        expect(instance.getSchemaKeys()).toEqual(['i18n.Global']);
     });
 
-    it('should return empty i18n keys array if specify empty string', function (done) {
+    it('should return empty i18n keys array if specify empty string', () => {
         const instance = validatorSchemaTranslations();
 
         instance.setSchemaKeys('');
 
-        expect(instance.getSchemaKeys()).equals([]);
-
-        done();
+        expect(instance.getSchemaKeys()).toEqual([]);
     });
 });
