@@ -1,43 +1,53 @@
-'use strict';
-
-const Glue = require('glue');
-const Hoek = require('hoek');
-const Url = require('url');
+const Glue = require('@hapi/glue');
+const _ = require('lodash');
 const manifest = require('./manifest');
 const logo = require('./lib/show-logo');
 
 require('colors');
 
-module.exports = (options, callback) => {
-    const config = manifest.get('/');
-    const parsedSecureUrl = Url.parse(options.dotStencilFile.storeUrl); //The url to a secure page (prompted as login page)
-    const parsedNormalUrl = Url.parse(options.dotStencilFile.normalStoreUrl); //The host url of the homepage;
+function buildManifest(srcManifest, options) {
+    const resManifest = _.cloneDeep(srcManifest);
+    const pluginsByName = resManifest.register.plugins;
 
-    callback = Hoek.nextTick(callback);
+    const parsedSecureUrl = new URL(options.dotStencilFile.storeUrl); // The url to a secure page (prompted as login page)
+    const parsedNormalUrl = new URL(options.dotStencilFile.normalStoreUrl); // The host url of the homepage;
+    const storeUrl = parsedSecureUrl.protocol + '//' + parsedSecureUrl.host;
 
-    config.connections[0].port = options.dotStencilFile.port;
-    config.plugins['./plugins/router/router.module'].storeUrl = parsedSecureUrl.protocol + '//' + parsedSecureUrl.host;
-    config.plugins['./plugins/router/router.module'].normalStoreUrl = parsedNormalUrl.protocol + '//' + parsedNormalUrl.host;
-    config.plugins['./plugins/router/router.module'].apiKey = options.dotStencilFile.apiKey;
-    config.plugins['./plugins/router/router.module'].port = options.dotStencilFile.port;
-    config.plugins['./plugins/router/router.module'].staplerUrl = options.dotStencilFile.staplerUrl;
-    config.plugins['./plugins/renderer/renderer.module'].useCache = options.useCache;
-    config.plugins['./plugins/renderer/renderer.module'].username = options.dotStencilFile.username;
-    config.plugins['./plugins/renderer/renderer.module'].token = options.dotStencilFile.token;
-    config.plugins['./plugins/renderer/renderer.module'].accessToken = options.dotStencilFile.accessToken;
-    config.plugins['./plugins/renderer/renderer.module'].customLayouts = options.dotStencilFile.customLayouts;
-    config.plugins['./plugins/renderer/renderer.module'].themePath = options.themePath;
-    config.plugins['./plugins/theme-assets/theme-assets.module'].themePath = options.themePath;
+    resManifest.server.port = options.dotStencilFile.port;
+    pluginsByName['./plugins/router/router.module'].storeUrl = storeUrl;
+    pluginsByName['./plugins/router/router.module'].normalStoreUrl =
+        parsedNormalUrl.protocol + '//' + parsedNormalUrl.host;
+    pluginsByName['./plugins/router/router.module'].apiKey = options.dotStencilFile.apiKey;
+    pluginsByName['./plugins/router/router.module'].port = options.dotStencilFile.port;
+    pluginsByName['./plugins/renderer/renderer.module'].useCache = options.useCache;
+    pluginsByName['./plugins/renderer/renderer.module'].username = options.dotStencilFile.username;
+    pluginsByName['./plugins/renderer/renderer.module'].token = options.dotStencilFile.token;
+    pluginsByName['./plugins/renderer/renderer.module'].accessToken =
+        options.dotStencilFile.accessToken;
+    pluginsByName['./plugins/renderer/renderer.module'].customLayouts =
+        options.dotStencilFile.customLayouts;
+    pluginsByName['./plugins/renderer/renderer.module'].themePath = options.themePath;
+    pluginsByName['./plugins/renderer/renderer.module'].storeUrl = storeUrl;
+    pluginsByName['./plugins/theme-assets/theme-assets.module'].themePath = options.themePath;
 
-    Glue.compose(config, {relativeTo: __dirname}, (err, server) => {
-        if (err) {
-            return callback(err);
-        }
+    resManifest.register.plugins = _.reduce(
+        pluginsByName,
+        (pluginsArr, opts, plugin) => [...pluginsArr, { plugin, options: opts }],
+        [],
+    );
 
-        server.start(() => {
-            // console.log(logo);
-            return callback(null, server);
-        });
+    return resManifest;
+}
 
-    });
-};
+async function create(options) {
+    const serverManifest = buildManifest(manifest.get('/'), options);
+
+    const server = await Glue.compose(serverManifest, { relativeTo: __dirname });
+    await server.start();
+
+    // console.log(logo);
+
+    return server;
+}
+
+module.exports = { create };
