@@ -1,28 +1,28 @@
 require('colors');
 
-var logo = '\n';
-logo += "                                     `+h\n".blue;
-logo += "                                   `+ddd\n".blue;
-logo += "                                 .oddddd\n".blue;
-logo += "                               .oddddddd\n".blue;
-logo += "                             -sddddddddd\n".blue;
-logo += "                          `-sddddddddddd\n".blue;
-logo += "                         -shdddddddddddd\n".blue;
-logo += "                         ...-:+ydddddddd\n".blue;
-logo += "                    `......`   `+ddddddd\n".blue;
-logo += "                    -ddddddh-    ddddddd\n".blue;
-logo += "                `   .yyyyyyo.  `+ddddddd\n".blue;
-logo += "              .o/    `````    :ydddddddd\n".blue;
-logo += "            -ohd+   `//////:` `.sddddddd\n".blue;
-logo += "          -sdddd+   -ddddddds   `hdddddd\n".blue;
-logo += "        :sdddddd+   .sssssso-   `ddddddd\n".blue;
-logo += "      :ydddddddd+              -yddddddd\n".blue;
-logo += "    /yddddddddddy+++++++++++oshddddddddd\n".blue;
-logo += " `/hdddddddddddddddddddddddddddddddddddd\n".blue;
-logo += "/hdddddddddddddddddddddddddddddddddddddd\n".blue;
-logo += "_________________________\n".white;
+let logo = '\n';
+logo += '                                     `+h\n'.blue;
+logo += '                                   `+ddd\n'.blue;
+logo += '                                 .oddddd\n'.blue;
+logo += '                               .oddddddd\n'.blue;
+logo += '                             -sddddddddd\n'.blue;
+logo += '                          `-sddddddddddd\n'.blue;
+logo += '                         -shdddddddddddd\n'.blue;
+logo += '                         ...-:+ydddddddd\n'.blue;
+logo += '                    `......`   `+ddddddd\n'.blue;
+logo += '                    -ddddddh-    ddddddd\n'.blue;
+logo += '                `   .yyyyyyo.  `+ddddddd\n'.blue;
+logo += '              .o/    `````    :ydddddddd\n'.blue;
+logo += '            -ohd+   `//////:` `.sddddddd\n'.blue;
+logo += '          -sdddd+   -ddddddds   `hdddddd\n'.blue;
+logo += '        :sdddddd+   .sssssso-   `ddddddd\n'.blue;
+logo += '      :ydddddddd+              -yddddddd\n'.blue;
+logo += '    /yddddddddddy+++++++++++oshddddddddd\n'.blue;
+logo += ' `/hdddddddddddddddddddddddddddddddddddd\n'.blue;
+logo += '/hdddddddddddddddddddddddddddddddddddddd\n'.blue;
+logo += '_________________________\n'.white;
 logo += '\n';
-logo += "BigCommerce Stencil \n".yellow;
+logo += 'BigCommerce Stencil \n'.yellow;
 logo += '_________________________\n'.white;
 
 module.exports = logo;
