@@ -1,103 +1,12 @@
 const _ = require('lodash');
 const Paper = require('@bigcommerce/stencil-paper');
-const internals = {};
-
-module.exports = function (data, assembler) {
-    this.respond = function (request, reply) {
-        const paper = new Paper(data.context.settings, data.context.theme_settings, assembler, "handlebars-v3");
-
-        // Set the environment to dev
-        data.context.in_development = true;
-        data.context.in_production = false;
-
-        paper.addDecorator(internals.makeDecorator(request, data.context));
-
-        // Plugins have the opportunity to add/modify the response by using decorators
-        _.each(request.app.decorators, function (decorator) {
-            paper.addDecorator(decorator);
-        });
-
-        const templatePath = internals.getTemplatePath(request, data);
-
-        paper.loadTheme(templatePath, data.acceptLanguage)
-            .then(() => {
-                if (request.query.debug === 'context') {
-                    return reply(data.context);
-                }
-            })
-            .catch(err => console.error(err.message.red))
-            .then(() => paper.renderTheme(templatePath, data))
-            .catch(err => console.error(err.message.red))
-            .then(output => {
-                const response = reply(output);
-                response.code(data.statusCode);
-                if (data.headers['set-cookie']) {
-                    response.header('set-cookie', data.headers['set-cookie']);
-                }
-                return response;
-            }).catch(err => console.error(err.message.red));
-    };
-};
-
-
-internals.getTemplatePath = function (request, data) {
-    var path = data.template_file;
-
-    if (request.headers['stencil-options']) {
-        var options = JSON.parse(request.headers['stencil-options']);
-
-        if (options['render_with'] && typeof options['render_with'] === 'string') {
-
-            path = options['render_with'].split(',');
-
-            path = _.map(path, function (path) {
-                return 'components/' + path;
-            });
-
-            if (path.length === 1) {
-                path = path[0];
-            }
-        }
-    }
-
-    return path;
-};
-
 
 /**
- * Output post-processing
+ * Escapes html entities
  *
- * @param request
- * @param context
+ * @param {string} html
  */
-internals.makeDecorator = function (request, context) {
-    return function(content) {
-        var regex,
-            debugBar;
-
-        if (context.settings) {
-            regex = new RegExp(internals.escapeRegex(context.settings.base_url), 'g');
-            content = content.replace(regex, '');
-
-            regex = new RegExp(internals.escapeRegex(context.settings.secure_base_url), 'g');
-            content = content.replace(regex, '');
-        }
-
-        if (request.query.debug === 'bar') {
-            debugBar = '<pre style="background-color:#EEE; word-wrap:break-word;">';
-            debugBar += internals.escapeHtml(JSON.stringify(context, null, 2)) + '</pre>';
-            regex = new RegExp('</body>');
-            content = content.replace(regex, debugBar + '\n</body>');
-        }
-
-        return content;
-    };
-};
-
-/**
- * Scape html entities
- */
-internals.escapeHtml = function (html) {
+const escapeHtml = (html) => {
     const charsToReplace = {
         '&': '&amp;',
         '<': '&lt;',
@@ -105,14 +14,159 @@ internals.escapeHtml = function (html) {
         '"': '&#34;',
     };
 
-    return html.replace(/[&<>"]/g, tag => charsToReplace[tag] || tag);
+    return html.replace(/[&<>"]/g, (tag) => charsToReplace[tag] || tag);
 };
 
 /**
  * Scape special characters for regular expression
  *
- * @param string
+ * @param {string} string
  */
-internals.escapeRegex = function (string) {
-    return string.replace(/[-\\^$*+?.()|[\]{}]/g, "\\$&");
+const escapeRegex = (string) => string.replace(/[-\\^$*+?.()|[\]{}]/g, '\\$&');
+
+/**
+ * Returns template path
+ *
+ * @param {string} request
+ * @param {Object} data
+ * @returns {string|string[]}
+ */
+const getTemplatePath = (request, data) => {
+    let paths = data.template_file;
+
+    if (request.headers['stencil-options']) {
+        const options = JSON.parse(request.headers['stencil-options']);
+
+        if (options.render_with && typeof options.render_with === 'string') {
+            const componentsToRender = options.render_with.split(',');
+
+            paths = componentsToRender.map((component) => `components/${component}`);
+
+            if (paths.length === 1) {
+                [paths] = paths;
+            }
+        }
+    }
+
+    return paths;
 };
+
+/**
+ * Returns boolean if this handlebars version is supported
+ *
+ * @param {string} version
+ */
+const isSupportedHandlebarsVersion = (version) =>
+    ['handlebars-v3', 'handlebars-v4'].includes(version);
+
+/**
+ * Node.js projects are using variables in format: handlebars-v3, handlebars-v4;
+ * Storefront and db are using format: handlebars_v3, handlebars_v4;
+ * This function converts _v3 to -v3
+ *
+ * @param {string} version
+ */
+const compatibilizeTemplateEngine = (version) => version.replace('_', '-');
+
+/**
+ * Output post-processing
+ *
+ * @param {Object} request
+ * @param {Object} context
+ */
+const makeDecorator = (request, context) => (content) => {
+    let updatedContent = content;
+
+    if (context.settings) {
+        const baseUrlregex = new RegExp(escapeRegex(context.settings.base_url), 'g');
+        updatedContent = content.replace(baseUrlregex, '');
+
+        const secureBaseUrlRegex = new RegExp(escapeRegex(context.settings.secure_base_url), 'g');
+        updatedContent = content.replace(secureBaseUrlRegex, '');
+    }
+
+    if (request.query.debug === 'bar') {
+        const contextStr = escapeHtml(JSON.stringify(context, null, 2));
+        const debugBar = `<pre style="background-color:#EEE; word-wrap:break-word;"> ${contextStr} </pre>`;
+        const bodyRegex = new RegExp('</body>');
+        updatedContent = content.replace(bodyRegex, `${debugBar}\n</body>`);
+    }
+
+    return updatedContent;
+};
+
+class PencilResponse {
+    /**
+     * @param {Object} data
+     * @param data.template_file
+     * @param data.templates
+     * @param data.remote
+     * @param data.remote_data
+     * @param data.context
+     * @param data.translations
+     * @param data.method
+     * @param data.acceptLanguage
+     * @param {{[string]: string[]}} data.headers
+     * @param data.statusCode
+     * @param data.renderedRegions
+     * @param assembler
+     */
+    constructor(data, assembler) {
+        this.data = data;
+        this.assembler = assembler;
+    }
+
+    async respond(request, h) {
+        const templateEngine = compatibilizeTemplateEngine(
+            this.data.context.template_engine || 'handlebars-v3',
+        );
+
+        if (!isSupportedHandlebarsVersion(templateEngine)) {
+            throw new Error(
+                'Provided Handlebars version is not supported! Please use:handlebars-v3, handlebars-v4',
+            );
+        }
+
+        const paper = new Paper(
+            this.data.context.settings,
+            this.data.context.theme_settings,
+            this.assembler,
+            templateEngine,
+        );
+        // Set the environment to dev
+        this.data.context.in_development = true;
+        this.data.context.in_production = false;
+
+        paper.addDecorator(makeDecorator(request, this.data.context));
+        paper.setContent(this.data.renderedRegions);
+
+        // Plugins have the opportunity to add/modify the response by using decorators
+        _.each(request.app.decorators, (decorator) => {
+            paper.addDecorator(decorator);
+        });
+
+        const templatePath = getTemplatePath(request, this.data);
+
+        if (request.query.debug === 'context') {
+            return this.data.context;
+        }
+
+        try {
+            await paper.loadTheme(templatePath, this.data.acceptLanguage);
+            const output = await paper.renderTheme(templatePath, this.data);
+
+            const response = h.response(output).code(this.data.statusCode);
+
+            if (this.data.headers['set-cookie']) {
+                response.header('set-cookie', this.data.headers['set-cookie']);
+            }
+
+            return response;
+        } catch (err) {
+            console.error(err.message.red);
+            return h.response().code(500);
+        }
+    }
+}
+
+module.exports = PencilResponse;
