@@ -1,109 +1,100 @@
-const Code = require('code');
-const Hapi = require('hapi');
-const Lab = require('lab');
-const lab = exports.lab = Lab.script();
-const describe = lab.describe;
-const expect = Code.expect;
-const it = lab.it;
+const Hapi = require('@hapi/hapi');
+const inert = require('@hapi/inert');
+const h2o2 = require('@hapi/h2o2');
 const router = require('./router.module');
 
 describe('Router', () => {
-    var server = new Hapi.Server();
+    const SERVER_OPTIONS = {
+        port: 3000,
+    };
+    const ROUTER_OPTIONS = {
+        storeUrl: 'https://store-abc124.mybigcommerce.com',
+        normalStoreUrl: 'http://s1234567890.mybigcommerce.com',
+        port: SERVER_OPTIONS.port,
+    };
+
+    const server = new Hapi.Server(SERVER_OPTIONS);
     const RendererPluginMock = {
-        register: function(server, options, next) {
-            server.expose('implementation', (request, reply) => reply('RendererHandlerFired'));
-
-            next();
+        register(_server) {
+            _server.expose('implementation', (request, h) => h.response('RendererHandlerFired'));
         },
-    };
-    const ThemeAssetsMock = {
-        register: function(server, options, next) {
-            server.expose('cssHandler', (request, reply) => reply('CssHandlerFired'));
-            server.expose('assetHandler', (request, reply) => reply('assetHandlerFired'));
-
-            next();
-        },
-    };
-
-    RendererPluginMock.register.attributes = {
         name: 'Renderer',
         version: '0.0.1',
     };
-
-    ThemeAssetsMock.register.attributes = {
+    const ThemeAssetsMock = {
+        register(_server) {
+            _server.expose('cssHandler', (request, h) => h.response('CssHandlerFired'));
+            _server.expose('assetHandler', (request, h) => h.response('assetHandlerFired'));
+        },
         name: 'ThemeAssets',
         version: '0.0.1',
     };
 
-    server.connection({
-        port: 3000,
-    });
-
-    lab.before(done => {
-        server.register([
+    beforeAll(async () => {
+        await server.register([
+            inert,
+            h2o2,
             RendererPluginMock,
             ThemeAssetsMock,
-            router,
-        ], err => {
-            expect(err).to.equal(undefined);
-            server.start(done);
-        });
+            { plugin: router, options: ROUTER_OPTIONS },
+        ]);
+
+        await server.start();
     });
 
-    lab.after(done => {
-        server.stop(done);
+    afterAll(async () => {
+        await server.stop();
     });
 
-    it('should call the Renderer handler', done => {
-        server.inject({
+    it('should call the Renderer handler', async () => {
+        const options = {
             method: 'GET',
             url: '/test',
-        }, response => {
-            expect(response.statusCode).to.equal(200);
-            expect(response.payload).to.equal('RendererHandlerFired');
+        };
 
-            done();
-        });
+        const response = await server.inject(options);
+
+        expect(response.statusCode).toEqual(200);
+        expect(response.payload).toEqual('RendererHandlerFired');
     });
 
-    it('should call the CSS handler', done => {
-        server.inject({
+    it('should call the CSS handler', async () => {
+        const options = {
             method: 'GET',
             url: '/stencil/123/css/file.css',
-        }, response => {
-            expect(response.statusCode).to.equal(200);
-            expect(response.payload).to.equal('CssHandlerFired');
+        };
 
-            done();
-        });
+        const response = await server.inject(options);
+
+        expect(response.statusCode).toEqual(200);
+        expect(response.payload).toEqual('CssHandlerFired');
     });
 
-    it('should call the assets handler', done => {
-        server.inject({
+    it('should call the assets handler', async () => {
+        const options = {
             method: 'GET',
             url: '/stencil/123/js/file.js',
-        }, response => {
-            expect(response.statusCode).to.equal(200);
-            expect(response.payload).to.equal('assetHandlerFired');
+        };
 
-            done();
-        });
+        const response = await server.inject(options);
+
+        expect(response.statusCode).toEqual(200);
+        expect(response.payload).toEqual('assetHandlerFired');
     });
 
-    it('should inject host and origin headers for GraphQL requests', done => {
-        server.inject({
+    it('should inject host and origin headers for GraphQL requests', async () => {
+        const options = {
             method: 'POST',
             url: '/graphql',
-            headers: { 'authorization': 'abc123' },
-        }, response => {
-            expect(response.request.payload.headers).to.include(
-                {
-                    authorization: 'abc123',
-                    origin: 'https://store-abc123.mybigcommerce.com',
-                    host: 'store-abc123.mybigcommerce.com',
-                },
-            );
-            done();
+            headers: { authorization: 'auth123' },
+        };
+
+        const response = await server.inject(options);
+
+        expect(response.request.payload.headers).toMatchObject({
+            authorization: 'auth123',
+            origin: 'https://store-abc124.mybigcommerce.com',
+            host: 'store-abc124.mybigcommerce.com',
         });
     });
 });
