@@ -1,47 +1,49 @@
 const path = require('path');
 
-/**
- * @param {Object} fs
- * @param {string} cwd
- * @param {Object} commandExecutor
- * @constructor
- */
-function ChangelogGenerator(fs, cwd, commandExecutor) {
+class ChangelogGenerator {
+    /**
+     * @param {Object} fs
+     * @param {string} cwd
+     * @param {Object} commandExecutor
+     * @constructor
+     */
+    constructor(fs, cwd, commandExecutor) {
+        this.fs = fs;
+        this.cwd = cwd;
+        this.commandExecutor = commandExecutor;
+    }
+
     /**
      * @param {ChangelogOptions} customOptions
      * @param {function(error: Error?): void} done
-     * @return {void}
+     * @returns {void}
      */
-    function generateChangelog(customOptions, done) {
-        const options = getOptions(customOptions);
-        commandExecutor.executeCommand('touch', ['package.json'], {}, () => {
-            commandExecutor.executeCommand('conventional-changelog', [], options, done);
+    generateChangelog(customOptions, done) {
+        const options = this._getOptions(customOptions);
+        this.commandExecutor.executeCommand('touch', ['package.json'], {}, () => {
+            this.commandExecutor.executeCommand('conventional-changelog', [], options, done);
         });
     }
 
     /**
      * @private
      * @param {ChangelogOptions} [customOptions={}]
-     * @return {ChangelogOptions}
+     * @returns {ChangelogOptions}
      */
-    function getOptions(customOptions) {
-        customOptions = customOptions ? customOptions : {};
-        const options = Object.assign({
+    _getOptions(customOptions = {}) {
+        const options = {
             config: customOptions.preset ? undefined : path.join(__dirname, 'default-config.js'),
-            infile: path.join(cwd, 'CHANGELOG.md'),
+            infile: path.join(this.cwd, 'CHANGELOG.md'),
             sameFile: true,
-        }, customOptions);
+            ...customOptions,
+        };
 
-        try {
-            fs.statSync(options.infile);
-        } catch(error) {
+        if (!this.fs.existsSync(options.infile)) {
             options.releaseCount = 0;
         }
 
         return options;
     }
-
-    this.generateChangelog = generateChangelog;
 }
 
 module.exports = ChangelogGenerator;
