@@ -1,65 +1,54 @@
+const { promisify } = require('util');
+const childProcess = require('child_process');
+const which = require('npm-which');
+
 const CommandExecutor = require('./command-executor');
-const Code = require('code');
-const Lab = require('lab');
-const lab = exports.lab = Lab.script();
-const describe = lab.describe;
-const sinon = require('sinon');
-const expect = Code.expect;
-const it = lab.it;
 
 describe('CommandExecutor', () => {
-    let commandExecutor;
-    let childProcess;
-    let spawnSpy;
-    let processMock;
-    let spawn;
-    let doneCallBack;
+    it('spawns a new process to run the given command', async () => {
+        const spawnSpy = jest.spyOn(childProcess, 'spawn').mockImplementation(jest.fn());
+        const commandExecutor = new CommandExecutor(childProcess);
+        const executeCommand = promisify(commandExecutor.executeCommand.bind(commandExecutor));
 
-    it('spawns a new process to run the given command', done => {
-        childProcess = require('child_process');
-        spawnSpy = sinon.spy(childProcess, "spawn");
-        commandExecutor = new CommandExecutor(childProcess);
+        try {
+            await executeCommand('jest', ['xyz.js'], { config: 'config.js' });
 
-        commandExecutor.executeCommand('jest', ['xyz.js'], { config: 'config.js' }, () => {
-            expect(spawnSpy.calledWith(
-                require('npm-which')(__dirname).sync('jest'),
-                ['--config', 'config.js', 'xyz.js'],
-                { stdio: 'inherit' },
-            )).to.equal(true);
+            // The executor will throw an error since '/xyz.js' doesn't exist, but we don't care,
+            //  just need to make sure that spawnSpy was called properly
+            // eslint-disable-next-line no-empty
+        } catch (err) {}
 
-            done();
-        });
+        expect(spawnSpy).toHaveBeenCalledWith(
+            which(__dirname).sync('jest'),
+            ['--config', 'config.js', 'xyz.js'],
+            { stdio: 'inherit' },
+        );
     });
 
-    it('returns undefined if the process exits with a successful exit code', done => {
-        processMock = {
-            on: function(event, callback) { return callback(0); },
+    // eslint-disable-next-line jest/expect-expect
+    it('resolves successfully if the process exits with a successful exit code', async () => {
+        const processMock = {
+            on(event, callback) {
+                return callback(0);
+            },
         };
+        const spawn = () => processMock;
+        const commandExecutor = new CommandExecutor({ spawn });
+        const executeCommand = promisify(commandExecutor.executeCommand.bind(commandExecutor));
 
-        spawn = function() { return processMock; };
-
-        doneCallBack = sinon.fake();
-        commandExecutor = new CommandExecutor({ spawn });
-        commandExecutor.executeCommand('jest', ['xyz.js'], { config: 'config.js' }, doneCallBack);
-
-        expect(doneCallBack.calledWith(undefined)).to.equal(true);
-
-        done();
+        await executeCommand('jest', ['xyz.js'], { config: 'config.js' });
     });
 
-    it('returns error if the process exits with an unsuccessful exit code', done => {
-        processMock = {
-            on: function(event, callback) { return callback(1); },
+    it('throws an error if the process exits with an unsuccessful exit code', async () => {
+        const processMock = {
+            on(event, callback) {
+                return callback(1);
+            },
         };
+        const spawn = () => processMock;
+        const commandExecutor = new CommandExecutor({ spawn });
+        const executeCommand = promisify(commandExecutor.executeCommand.bind(commandExecutor));
 
-        spawn = function() { return processMock; };
-
-        doneCallBack = sinon.fake();
-        commandExecutor = new CommandExecutor({ spawn });
-        commandExecutor.executeCommand('jest', ['xyz.js'], { config: 'config.js' }, doneCallBack);
-
-        expect(doneCallBack.calledWith(undefined)).to.equal(false);
-
-        done();
+        await expect(executeCommand('jest', ['xyz.js'], { config: 'config.js' })).rejects.toThrow();
     });
 });
