@@ -1,13 +1,15 @@
 /**
  * @param {Object} commit
- * @return {Object}
+ * @returns {Object}
  */
 function transform(commit) {
     if (/^Releasing \d+\.\d+\.\d+/.test(commit.header)) {
         return false;
     }
 
+    // eslint-disable-next-line no-param-reassign
     commit.hash = commit.hash.slice(0, 7);
+    // eslint-disable-next-line no-param-reassign
     commit.references = [];
 
     return commit;
